% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{as.data.frame,BASiCS_ResultsDE-method}
\alias{as.data.frame,BASiCS_ResultsDE-method}
\alias{as.data.frame,BASiCS_ResultDE-method}
\alias{as.data.frame,BASiCS_ResultVG-method}
\title{Converting BASiCS results objects to data.frames}
\usage{
\S4method{as.data.frame}{BASiCS_ResultsDE}(x, Parameter, Filter = TRUE, ProbThreshold = NULL)

\S4method{as.data.frame}{BASiCS_ResultDE}(x, Filter = TRUE, ProbThreshold = NULL)

\S4method{as.data.frame}{BASiCS_ResultVG}(x, Filter = TRUE, ProbThreshold = NULL)
}
\arguments{
\item{x}{An object of class \linkS4class{BASiCS_ResultVG}, 
\linkS4class{BASiCS_ResultDE}, or \linkS4class{BASiCS_ResultsDE}.}

\item{Parameter}{For \linkS4class{BASiCS_ResultsDE} objects only. 
Character scalar specifying which table of results to 
output. Available options are "Mean", (mu, mean expression),
"Disp" (delta, overdispersion) and "ResDisp" 
(epsilon, residual overdispersion).}

\item{Filter}{Logical scalar. If \code{TRUE}, output only entries
corresponding to genes that pass the decision rule used in the probabilistic
test.}

\item{ProbThreshold}{Only used if \code{Filter=TRUE}.
Numeric scalar specifying the probability threshold to be used when filtering
genes. Default is to use the threshold used in the original decision rule
when the test was performed.}
}
\description{
Converting BASiCS results objects to data.frames
}
