\name{BASiCS-defunct}
\alias{BASiCS-defunct}
\alias{BASiCS_D_TestDE}
\title{Defunct functions in package \sQuote{BASiCS}}

\description{
  The functions listed here are no longer part of \code{BASiCS}.
}

\section{usage}{

  ## Removed
  \itemize{
    \item BASiCS_D_TestDE()
  }
}

\details{

  ## Removed
  \itemize{
    \item{
      \code{BASiCS_D_TestDE} has been replaced by \code{BASiCS_TestDE}.
    }
  }
}

\author{
 Catalina A. Vallejos \email{cnvallej@uc.cl}
}

\seealso{
  \itemize{
    \item \code{\link{BASiCS_TestDE}}
  }
}


