\name{BASiCS-deprecated}
\alias{BASiCS-deprecated}

\title{Deprecated functions in package \sQuote{BASiCS}}

\description{
  These functions are provided for compatibility with older versions
  of \sQuote{BASiCS} only, and will be defunct at the next release.
}

\details{

  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
  \itemize{

    \item{BASiCS_D_TestDE: \code{\link{BASiCS_TestDE}}}

  }
}

