% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASiCS_CalculateSpikeIns.R
\name{BASiCS_CalculateSpikeIns}
\alias{BASiCS_CalculateSpikeIns}
\title{Convert concentration in moles per microlitre to molecule counts}
\usage{
BASiCS_CalculateSpikeIns(ERCCConcentration, DilutionFactor, VolumePerCell)
}
\arguments{
\item{ERCCConcentration}{A vector of concentrations, in moles per litre,
of the spike-in genes in the mixture used.}

\item{DilutionFactor}{The dilution factor applied to the spike-in mixture.
e.g., 1μl per 50ml would be a 1/50000 \code{DilutionFactor}.}

\item{VolumePerCell}{The volume of spike-in mixture added to each well, or to
each cell.}
}
\value{
The molecule counts per well, or per cell, based on the input
parameters.
}
\description{
Convert concentration in moles per microlitre to molecule counts
}
