% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FunctionsPart1.R
\name{displayChainMu}
\alias{displayChainDelta}
\alias{displayChainDelta,BASiCS_Chain-accessors}
\alias{displayChainMu}
\alias{displayChainMu,BASiCS_Chain-accessors}
\alias{displayChainNu}
\alias{displayChainNu,BASiCS_Chain-accessors}
\alias{displayChainPhi}
\alias{displayChainPhi,BASiCS_Chain-accessors}
\alias{displayChainS}
\alias{displayChainS,BASiCS_Chain-accessors}
\alias{displayChainTheta}
\alias{displayChainTheta,BASiCS_Chain-accessors}
\title{Accessors for the slots of a BASiCS_Chain object}
\usage{
displayChainMu(object)

displayChainDelta(object)

displayChainPhi(object)

displayChainS(object)

displayChainNu(object)

displayChainTheta(object)
}
\arguments{
\item{object}{an object of class \code{\link[BASiCS]{BASiCS_Chain-class}}}
}
\value{
\describe{
\item{\code{displayChainMu}}{The \code{mu} slot of an object of class \code{\link[BASiCS]{BASiCS_Chain-class}}.}
\item{\code{displayChainDelta}}{The \code{delta} slot of an object of class \code{\link[BASiCS]{BASiCS_Chain-class}}.}
\item{\code{displayChainPhi}}{The \code{phi} slot of an object of class \code{\link[BASiCS]{BASiCS_Chain-class}}.}
\item{\code{displayChainS}}{The \code{s} slot of an object of class \code{\link[BASiCS]{BASiCS_Chain-class}}.}
\item{\code{displayChainNu}}{The \code{nu} slot of an object of class \code{\link[BASiCS]{BASiCS_Chain-class}}.}
\item{\code{displayChainTheta}}{The \code{theta} slot of an object of class \code{\link[BASiCS]{BASiCS_Chain-class}}.}
}
}
\description{
Accessors for the slots of a \code{\link[BASiCS]{BASiCS_Chain-class}}
}
\examples{
# See
help(BASiCS_MCMC)
}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}
\references{
Vallejos, Marioni and Richardson (2015). Bayesian Analysis of Single-Cell Sequencing data.
}
\seealso{
\code{\link[BASiCS]{BASiCS_Chain-class}}
}

