% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{BASiCS_Chain-class}
\alias{BASiCS_Chain}
\alias{BASiCS_Chain,BASiCS_Chain-class}
\alias{BASiCS_Chain-class}
\title{The BASiCS_Chain class}
\description{
Container of an MCMC sample of the BASiCS' model parameters (see Vallejos et al, 2015) as generated
by the function \code{\link[BASiCS]{BASiCS_MCMC}}.
}
\section{Slots}{

\describe{
\item{\code{mu}}{MCMC chain for gene-specific expression levels \eqn{\mu[i]}, defined as true input molecules in case of technical genes 
(matrix with \code{q} columns, technical genes located at the end of the matrix, all elements must be positive numbers)}

\item{\code{delta}}{MCMC chain for gene-specific biological cell-to-cell heterogeneity hyper-parameters \eqn{\delta[i]}, biological genes only 
(matrix with \code{q.bio} columns, all elements must be positive numbers)}

\item{\code{phi}}{MCMC chain for cell-specific mRNA content normalising constants \eqn{\phi[j]}
(matrix with \code{n} columns, all elements must be positive numbers and the sum of its elements must be equal to \code{n})}

\item{\code{s}}{MCMC chain for cell-specific capture efficiency (or amplification biases if not using UMI based counts) normalising constants \eqn{s[j]}
(matrix with \code{n} columns, all elements must be positive numbers)}

\item{\code{nu}}{MCMC chain for cell-specific random effects \eqn{\nu[j]}
(matrix with \code{n} columns, all elements must be positive numbers)}

\item{\code{theta}}{MCMC chain for technical variability hyper-parameter(s) \eqn{\theta} (matrix, all elements must be positive, each colum 
represents 1 batch)}
}}
\examples{

# A BASiCS_Chain object created by the BASiCS_MCMC function.
Data = makeExampleBASiCS_Data()
MCMC_Output <- BASiCS_MCMC(Data, N = 100, Thin = 2, Burn = 2)

}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}
\references{
Vallejos, Marioni and Richardson (2015). Bayesian Analysis of Single-Cell Sequencing data. PLoS Computational Biology.
}

