% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\docType{methods}
\name{BASiCS_Chain-methods}
\alias{BASiCS_Chain-methods}
\alias{show,BASiCS_Chain-method}
\alias{BASiCS_Chain-methods}
\alias{updateObject,BASiCS_Chain-method}
\title{'show' method for BASiCS_Chain objects}
\usage{
\S4method{show}{BASiCS_Chain}(object)

\S4method{updateObject}{BASiCS_Chain}(object, ..., verbose = FALSE)
}
\arguments{
\item{object}{A \code{\linkS4class{BASiCS_Chain}} object.}

\item{...}{Additional arguments of \code{\link[BiocGenerics]{updateObject}}
generic method. Not used within BASiCS.}

\item{verbose}{Additional argument of \code{\link[BiocGenerics]{updateObject}}
generic method. Not used within BASiCS.}
}
\value{
Prints a summary of the properties of \code{object}.

Returns an updated \code{\linkS4class{BASiCS_Chain}} object that 
contains all model parameters in a single slot object (list).
}
\description{
'show' method for \code{\linkS4class{BASiCS_Chain}} objects.

'updateObject' method for \code{\linkS4class{BASiCS_Chain}} 
objects. It is used to convert outdated \code{\linkS4class{BASiCS_Chain}}
objects into a version that is compatible with the Bioconductor release
of BASiCS. Do not use this method is \code{\linkS4class{BASiCS_Chain}} 
already contains a \code{parameters} slot.
}
\examples{

Data <- makeExampleBASiCS_Data()
Chain <- BASiCS_MCMC(Data, N = 50, Thin = 2, Burn = 2)


# Not run
# New_Chain <- updateObject(Old_Chain)

}
\references{
Vallejos, Marioni and Richardson (2015). PLoS Computational Biology.

Vallejos, Marioni and Richardson (2015). PLoS Computational Biology.
}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}

Catalina A. Vallejos \email{cnvallej@uc.cl}
}
