% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{BASiCS_D_Chain-class}
\alias{BASiCS_D_Chain-class}
\alias{BASiCS_D_Chain}
\alias{BASiCS_D_Chain,BASiCS_D_Chain-class}
\title{The BASiCS_D_Chain class}
\description{
Container of an MCMC sample of the BASiCS model parameters when comparing 2 populations of cells. 
For each population, these are independently generated by the function \code{\link[BASiCS]{BASiCS_MCMC}} and 
afterwards combined through the function \code{\link[BASiCS]{CombineBASiCS_Chain}} .
}
\section{Slots}{

\describe{
\item{\code{muTest}}{MCMC chain for gene-specific expression levels \eqn{\mu[i]} in test group, defined as true input molecules in case of technical genes 
(matrix with \code{q.bio} columns, all elements must be positive numbers)}

\item{\code{muRef}}{MCMC chain for gene-specific expression levels \eqn{\mu[i]} in reference group, defined as true input molecules in case of technical genes 
(matrix with \code{q.bio} columns, all elements must be positive numbers)}

\item{\code{deltaTest}}{MCMC chain for gene-specific biological cell-to-cell heterogeneity hyper-parameters \eqn{\delta[i]} 
in test group, biological genes only (matrix with \code{q} columns, all elements must be positive numbers)}

\item{\code{deltaRef}}{MCMC chain for gene-specific biological cell-to-cell heterogeneity hyper-parameters \eqn{\delta[i]} in reference group, biological genes only 
(matrix with \code{q} columns, all elements must be positive numbers)}

\item{\code{phi}}{MCMC chain for cell-specific mRNA content normalising constants \eqn{\phi[j]}
(matrix with \code{n} columns, all elements must be positive numbers and the sum of its elements must be equal to \code{n})}

\item{\code{s}}{MCMC chain for cell-specific capture efficiency (or amplification biases if not using UMI based counts) normalising constants \eqn{s[j]}
(matrix with \code{n} columns, all elements must be positive numbers)}

\item{\code{nu}}{MCMC chain for cell-specific random effects \eqn{\nu[j]}
(matrix with \code{n} columns, all elements must be positive numbers)}

\item{\code{thetaTest}}{MCMC chain for technical variability hyper-parameter \eqn{\theta[test]} in the test sample (vector, all elements must be positive)}

\item{\code{thetaRef}}{MCMC chain for technical variability hyper-parameter \eqn{\theta[ref]} in the reference sample (vector, all elements must be positive)}

\item{\code{offset}}{Offset value to capture global changes in mRNA content. Default value = 1 (when no offset correction has been performed)}
}}

\examples{

# See vignette

}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}
