% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\docType{methods}
\name{BASiCS_D_Chain-methods}
\alias{BASiCS_D_Chain-methods}
\alias{show,BASiCS_D_Chain-method}
\title{'show' method for BASiCS_D_Chain objects}
\usage{
\S4method{show}{BASiCS_D_Chain}(object)
}
\arguments{
\item{object}{A \code{BASiCS_D_Chain} object.}
}
\value{
Prints a summary of the properties of \code{object}.
}
\description{
'show' method for \code{\link[BASiCS]{BASiCS_D_Chain-class}} objects.
}
\examples{

Data = makeExampleBASiCS_D_Data()
#MCMC_Output <- BASiCS_D_MCMC(Data, N = 50, Thin = 2, Burn = 2)

}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}

