% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{BASiCS_D_Summary-class}
\alias{BASiCS_D_Summary-class}
\alias{BASiCS_D_Summary,BASiCS_D_Summary-class}
\title{The BASiCS_D_Summary class contains posterior medians and the limits of 
the highest posterior density (HPD) interval for all BASiCS model parameters}
\description{
Container of a summary of a \code{\link[BASiCS]{BASiCS_D_Chain-class}} object.
}
\section{Slots}{

\describe{
\item{\code{muTest}}{Posterior medians (first column), lower (second column) and upper (third column) 
limits of HPD interval for gene-specific expression levels \eqn{\mu[i]} in test group.}

\item{\code{muRef}}{Posterior medians (first column), lower (second column) and upper (third column) 
limits of HPD interval for gene-specific expression levels \eqn{\mu[i]} in reference group.}

\item{\code{deltaTest}}{Posterior medians (first column), lower (second column) and upper (third column) 
limits of HPD interval for gene-specific biological cell-to-cell heterogeneity hyper-parameters \eqn{\delta[i]} in test group}

\item{\code{deltaRef}}{Posterior medians (first column), lower (second column) and upper (third column) 
limits of HPD interval for gene-specific biological cell-to-cell heterogeneity hyper-parameters \eqn{\delta[i]} in reference group}

\item{\code{phi}}{Posterior medians (first column), lower (second column) and upper (third column) 
limits of HPD interval for cell-specific mRNA content normalising constants \eqn{\phi[j]}}

\item{\code{s}}{Posterior medians (first column), lower (second column) and upper (third column) 
limits of HPD interval for cell-specific capture efficiency (or amplification biases if not using UMI based counts) normalising constants \eqn{s[j]}}

\item{\code{of}}{HPD interval for Posterior medians (first column), lower (second column) and upper (third column) 
limits of HPD interval for cell-specific random effects \eqn{\nu[j]}}

\item{\code{thetaTest}}{Posterior median (first column), lower (second column) and upper (third column) 
limits of HPD interval for technical variability hyper-parameter \eqn{\theta[test]} for test sample (vector, all elements must be positive)}

\item{\code{thetaRef}}{Posterior median (first column), lower (second column) and upper (third column) 
limits of HPD interval for technical variability hyper-parameter \eqn{\theta[ref]} for reference sample (vector, all elements must be positive)}

\item{\code{offset}}{Offset value to capture global changes in mRNA content. Default value = 1 (when no offset correction has been performed)}

\item{\code{probHPD}}{Posterior probability contained in the HPD interval. Must be a value between 0 and 1.}
}}

\examples{

# see help(BASiCS_D_MCMC)

}
