% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASiCS_D_TestDE.R
\name{BASiCS_D_TestDE}
\alias{BASiCS_D_TestDE}
\title{Detection of genes with changes in expression.}
\usage{
BASiCS_D_TestDE(...)
}
\arguments{
\item{...}{Optional parameters.}
}
\value{
This function is no longer in use and will be removed from future releases of \code{BASiCS}. Please use \code{\link[BASiCS]{BASiCS_TestDE}} instead.
}
\description{
This function is no longer in use and will be removed from future releases of \code{BASiCS}. Please use \code{\link[BASiCS]{BASiCS_TestDE}} instead.
}
\details{
This function is no longer in use and will be removed from future releases of \code{BASiCS}.
}
\examples{

# This function is no longer in use and will be removed in future releases of \\code{BASiCS}.

}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}
}
