% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{BASiCS_Data-class}
\alias{BASiCS_Data,BASiCS_Data-class}
\alias{BASiCS_Data-class}
\title{The BASiCS_Data class}
\description{
Container of expression counts from single-cell 
sequencing experiments in the format required by BASiCS (see Vallejos et al 2015).
}
\section{Slots}{

\describe{
\item{\code{Counts}}{Matrix of dimensions \code{q} times \code{n} whose elements corresponds to the raw expression counts. 
First \code{q.bio} rows correspond to biological genes. Last \code{q-q.bio} rows correspond to technical spike-in genes.}

\item{\code{Tech}}{Logical vector of length \code{q}. If \code{Tech = F} the gene is biological; otherwise the gene is spike-in.}

\item{\code{SpikeInput}}{Vector of length \code{q-q.bio} whose elements indicate the input number of molecules for the spike-in genes (amount per cell).}

\item{\code{GeneNames}}{Vector of length \code{q} containing gene names. Default value: \code{GeneNames = paste("Gene", 1:q)}, 
with numbering order as in the input dataset.}

\item{\code{BatchInfo}}{Vector of lenght \code{n} containing batch indicators.}
}}
\examples{

# Expression counts
set.seed(1)
Counts = Counts = matrix(rpois(50*10, 2), ncol = 10)
rownames(Counts) <- c(paste0("Gene", 1:40), paste0("Spike", 1:10))

# Technical information 
Tech = c(rep(FALSE,40),rep(TRUE,10))

# Spikes input number of molecules 
set.seed(2)
SpikeInfo <- data.frame(gene=rownames(Counts)[Tech],amount=rgamma(10,1,1))

# Creating a BASiCS_Data object (no batch effect)
Data = newBASiCS_Data(Counts, Tech, SpikeInfo)

head(counts(Data))
dim(counts(Data, type="biological"))
dim(counts(Data, type="technical"))
displayTechIndicator(Data)
displaySpikeInput(Data)
displayBatchInfo(Data)

}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}
\references{
Vallejos, Marioni and Richardson (2015). Bayesian Analysis of Single-Cell Sequencing data. PLoS Computational Biology.
}
\seealso{
\code{\link[BASiCS]{BASiCS_Data-methods}}, \code{\link[BASiCS]{BASiCS_MCMC}}.
}

