% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{BASiCS_Data-methods}
\alias{BASiCS_Data-methods}
\alias{counts}
\alias{counts,BASiCS_Data-method}
\alias{displaySpikeInput}
\alias{displaySpikeInput,BASiCS_Data-method}
\alias{displayTechIndicator}
\alias{displayTechIndicator,BASiCS_Data-method}
\alias{show}
\alias{show,BASiCS_Data-method}
\title{S4 methods for BASiCS_Data objects}
\usage{
\S4method{show}{BASiCS_Data}(object)

\S4method{counts}{BASiCS_Data}(object, type = "all")

\S4method{displaySpikeInput}{BASiCS_Data}(object)

\S4method{displayTechIndicator}{BASiCS_Data}(object)
}
\arguments{
\item{object}{A \code{BASiCS_Data} object.}

\item{type}{Only required for \code{counts} method. A string indicating which genes are returned. If \code{all}, returns expression counts for all genes.
If \code{biological}, returns expression counts for biological genes only. If \code{technical}, returns expression
counts for technical (spike-in) genes only.}
}
\value{
\describe{
\item{\code{show}}{Prints a summary of the properties of \code{object}.}

\item{\code{counts}}{If \code{type = "all"}, returns the \code{Counts} slot of \code{object}.
If \code{type = "biological"} or \code{type = "spike-in"}, only biological or technical (spike-in) counts are returned, respectively.}

\item{\code{displayTechIndicator}}{Returns \code{Tech} slot of \code{object}.}

\item{\code{displaySpikeInput}}{Returns \code{SpikeInput} slot of \code{object}.}
}
}
\description{
S4 methods for \code{\link[BASiCS]{BASiCS_Data-class}} objects.
}
\examples{
Data = makeExampleBASiCS_Data()
show(Data)
head(counts(Data))
dim(counts(Data, type="biological"))
dim(counts(Data, type="technical"))
displayTechIndicator(Data)
displaySpikeInput(Data)
}
\seealso{
\code{\link[BASiCS]{BASiCS_Data-class}}, \code{\link[BASiCS]{BASiCS_MCMC_Start}}, \code{\link[BASiCS]{BASiCS_MCMC}}.
}

