% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASiCS_PlotDE.R
\docType{methods}
\name{BASiCS_DEPlots,BASiCS_ResultsDE-method}
\alias{BASiCS_DEPlots,BASiCS_ResultsDE-method}
\title{Produce plots assessing differential expression results}
\usage{
\S4method{BASiCS_DEPlots}{BASiCS_ResultsDE}(object, Which = c("MAPlot",
  "VolcanoPlot", "GridPlot"), ...)
}
\arguments{
\item{Which}{}

\item{...}{Passed to methods}

\item{ResultsDE}{A BASiCS_ResultsDE object}
}
\description{
Produce plots assessing differential expression results
}
