% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASiCS_diagPlot.R
\name{BASiCS_diagPlot}
\alias{BASiCS_diagPlot}
\title{Create diagnostic plots of MCMC parameters}
\usage{
BASiCS_diagPlot(object, Param = "mu", x = NULL, y = NULL,
  LogX = isTRUE(x \%in\% c("mu", "delta")), LogY = isTRUE(y \%in\%
  c("mu", "delta")), na.rm = FALSE)
}
\arguments{
\item{object}{an object of class \code{\linkS4class{BASiCS_Summary}}}

\item{Param}{Optional name of a chain parameter to restrict the histogram;
if not supplied, all parameters will be assessed.
Possible values: \code{'mu'}, \code{'delta'}, \code{'phi'},
\code{'s'}, \code{'nu'}, \code{'theta'}, \code{'beta'},
\code{'sigma2'} and \code{'epsilon'}.}

\item{x, y}{Optional MCMC parameter values to be plotted on the x or y axis, 
respectively. If neither is supplied, Param will be plotted on the x axis
and \code{coda::effectiveSize(Param)} will be plotted on the y axis as
a density plot.}

\item{LogX, LogY}{A boolean value indicating whether to use a log10
transformation for the x or y axis, respectively.}

\item{na.rm}{Logical value indicating whether NA values should be removed
before calculating effective sample size.}
}
\value{
A ggplot object.
}
\description{
Plot parameter values and effective sample size.
See \link[coda]{effectiveSize}
for more details on this diagnostic measure.
}
\examples{

# Built-in simulated dataset
Data <- makeExampleBASiCS_Data()
# To analyse real data, please refer to the instructions in:
# https://github.com/catavallejos/BASiCS/wiki/2.-Input-preparation

# Only a short run of the MCMC algorithm for illustration purposes
# Longer runs migth be required to reach convergence
Chain <- BASiCS_MCMC(Data, N = 10, Thin = 2, Burn = 4, Regression = FALSE,
                     PrintProgress = FALSE, WithSpikes = TRUE)
# Plot effective sample size for mu (y) against mu, coloured by density
BASiCS_diagPlot(Chain)
# Effective sample size as colour, mu as x, delta as y.
BASiCS_diagPlot(Chain, x = "mu", y = "delta")

}
\seealso{
\code{\linkS4class{BASiCS_Chain}}
}
\author{
Alan O'Callaghan \email{a.b.ocallaghan@sms.ed.ac.uk}
}
