% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{BASiCS_DV_Chain-class}
\alias{BASiCS_DV_Chain}
\alias{BASiCS_DV_Chain,BASiCS_DV_Chain-class}
\alias{BASiCS_DV_Chain-class}
\title{The BASiCS_DV_Chain class}
\description{
Container of an MCMC sample of the BASiCSDV' model parameters as generated
by the function \code{\link[BASiCSDV]{BASiCS_DV_MCMC}}.
}
\section{Slots}{

\describe{
\item{\code{muTest}}{MCMC chain for gene-specific expression levels \eqn{\mu[i]} in test group, defined as true input molecules in case of technical genes 
(matrix with \code{q.bio} columns, all elements must be positive numbers)}

\item{\code{muRef}}{MCMC chain for gene-specific expression levels \eqn{\mu[i]} in reference group, defined as true input molecules in case of technical genes 
(matrix with \code{q.bio} columns, all elements must be positive numbers)}

\item{\code{deltaTest}}{MCMC chain for gene-specific biological cell-to-cell heterogeneity hyper-parameters \eqn{\delta[i]} 
in test group, biological genes only (matrix with \code{q} columns, all elements must be positive numbers)}

\item{\code{deltaRef}}{MCMC chain for gene-specific biological cell-to-cell heterogeneity hyper-parameters \eqn{\delta[i]} in reference group, biological genes only 
(matrix with \code{q} columns, all elements must be positive numbers)}

\item{\code{phi}}{MCMC chain for cell-specific mRNA content normalising constants \eqn{\phi[j]}
(matrix with \code{n} columns, all elements must be positive numbers and the sum of its elements must be equal to \code{n})}

\item{\code{s}}{MCMC chain for cell-specific capture efficiency (or amplification biases if not using UMI based counts) normalising constants \eqn{s[j]}
(matrix with \code{n} columns, all elements must be positive numbers)}

\item{\code{nu}}{MCMC chain for cell-specific random effects \eqn{\nu[j]}
(matrix with \code{n} columns, all elements must be positive numbers)}

\item{\code{thetaTest}}{MCMC chain for technical variability hyper-parameter \eqn{\theta[test]} in the test sample (vector, all elements must be positive)}

\item{\code{thetaRef}}{MCMC chain for technical variability hyper-parameter \eqn{\theta[ref]} in the reference sample (vector, all elements must be positive)}
}}
\examples{

# A BASiCS_Chain object created by the BASiCS_MCMC function.
Data = makeExampleBASiCS_DV_Data()
# MCMC_Output <- BASiCS_DV_MCMC(Data, N = 100, Thin = 2, Burn = 2)

}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}

