% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{BASiCS_DV_Data-class}
\alias{BASiCS_DV_Data,BASiCS_DV_Data-class}
\alias{BASiCS_DV_Data-class}
\title{The BASiCS_DV_Data class}
\description{
Container of expression counts from single-cell
sequencing experiments in the format required to perform the differential
expression and differential variability tests using BASiCS_DV.
}
\section{Slots}{

\describe{
\item{\code{CountsTest}}{Matrix of dimensions \code{q} times \code{n_test} whose elements corresponds to the raw expression counts in the test sample.
First \code{q.bio} rows correspond to biological genes. Last \code{q-q.bio} rows correspond to technical spike-in genes.}

\item{\code{CountsRef}}{Matrix of dimensions \code{q} times \code{n_ref} whose elements corresponds to the raw expression counts in the reference sample.
First \code{q.bio} rows correspond to biological genes. Last \code{q-q.bio} rows correspond to technical spike-in genes.}

\item{\code{Tech}}{Logical vector of length \code{q}. If \code{Tech = F} the gene is biological; otherwise the gene is spike-in.}

\item{\code{SpikeInputTest}}{Vector of length \code{q-q.bio} whose elements indicate the input number of molecules for the spike-in genes in the test sample (amount per cell).}

\item{\code{SpikeInputRef}}{Vector of length \code{q-q.bio} whose elements indicate the input number of molecules for the spike-in genes in the reference sample (amount per cell).}
}}
\examples{
CountsTest = matrix(rpois(100*5, 2), ncol = 5)
CountsRef = matrix(rpois(100*5, 2), ncol = 5)
Tech = c(rep(FALSE,80),rep(TRUE,20))
SpikeInputTest = rgamma(20,1,1)
SpikeInputRef = rgamma(20,1,1)
Data = newBASiCS_DV_Data(CountsTest, CountsRef, Tech, SpikeInputTest, SpikeInputRef)

#head(counts(Data))
#dim(counts(Data, sample = "test", type="biological"))
#dim(counts(Data, sample = "reference", type="technical"))
#displayTechIndicator(Data)
#displaySpikeInput(Data)
}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}
\references{
Vallejos, Marioni and Richardson (2015). Bayesian Analysis of Single-Cell Sequencing data.
}

