% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Methods.R
\docType{methods}
\name{BASiCS_DV_Data-methods}
\alias{BASiCS_DV_Data-methods}
\alias{counts}
\alias{counts,BASiCS_DV_Data-method}
\alias{displaySpikeInput}
\alias{displaySpikeInput,BASiCS_DV_Data-method}
\alias{displayTechIndicator}
\alias{displayTechIndicator,BASiCS_DV_Data-method}
\alias{show,BASiCS_DV_Data-method}
\title{S4 methods for BASiCS_DV_Data objects}
\usage{
\S4method{show}{BASiCS_DV_Data}(object)

\S4method{counts}{BASiCS_DV_Data}(object, type = "all", samples = "test")

\S4method{displaySpikeInput}{BASiCS_DV_Data}(object, samples = "test")

\S4method{displayTechIndicator}{BASiCS_DV_Data}(object)
}
\arguments{
\item{object}{A \code{BASiCS_DV_Data} object.}

\item{type}{Only required for \code{counts} method. A string indicating which genes must be returned. Valid values: \code{"all", "biological"} and \code{"technical"}}

\item{samples}{A string indicating which samples must be returned. Valid values: \code{"test"} and \code{"reference"}.}
}
\value{
\describe{
\item{\code{show}}{Prints a summary of the properties of \code{object}.}

\item{\code{counts}}{\describe{
 \item{\code{if(type = "all")}}{Returns the \code{CountsTest} or \code{CountsRef} slot
 of \code{object} (depending on the value of \code{sample}).}
 \item{\code{if(type = "biological")}}{Returns the \code{CountsTest} or \code{CountsRef} slot of \code{object}
 (depending on the value of \code{sample}), biological genes only.}
 \item{\code{if(type = "technical")}}{Returns the \code{CountsTest} or \code{CountsRef} slot of \code{object}
 (depending on the value of \code{sample}), tehcnical genes only.}
 \item{\code{if(samples == "test")}}{Returns counts in the test sample.}
 \item{\code{if(samples == "reference")}}{Returns counts in the reference sample.}
 }}

\item{\code{displayTechIndicator}}{Returns \code{Tech} slot of \code{object}.}

\item{\code{displaySpikeInput}}{\describe{
\item{\code{if(samples == "test")}}{Returns \code{SpikeInputTest} slot of \code{object}.}
\item{\code{if(samples == "reference")}}{Returns \code{SpikeInputRef} slot of \code{object}.}}}}
}
\description{
S4 methods for \code{\link[BASiCSDV]{BASiCS_DV_Data-class}} objects.
}
\examples{
Data = makeExampleBASiCS_DV_Data()
show(Data)
head(counts(Data, samples = "test"))
dim(counts(Data, type="biological", samples = "test"))
dim(counts(Data, type="technical", samples = "reference"))
displayTechIndicator(Data)
displaySpikeInput(Data, samples = "test")
}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}
\seealso{
\code{\link[BASiCSDV]{BASiCS_DV_Data-class}}
}

