% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{BASiCS_DV_Summary-class}
\alias{BASiCS_DV_Summary,BASiCS_DV_Summary-class}
\alias{BASiCS_DV_Summary-class}
\title{The BASiCS_DV_Summary class}
\description{
Container of a summary of a \code{\link[BASiCSDV]{BASiCS_DV_Chain-class}} object.  
In each slot, first column contains posterior medians, second column contains the lower limits of an high posterior
density interval and third column contains the upper limits of high posterior density intervals.
}
\section{Slots}{

\describe{
\item{\code{muTest}}{Posterior medians (first column), lower (second column) and upper (third column) 
limits of gene-specific expression levels \eqn{\mu[i]} in test group.}

\item{\code{muRef}}{Posterior medians (first column), lower (second column) and upper (third column) 
limits of gene-specific expression levels \eqn{\mu[i]} in reference group.}

\item{\code{deltaTest}}{Posterior medians (first column), lower (second column) and upper (third column) 
limits of gene-specific biological cell-to-cell heterogeneity hyper-parameters \eqn{\delta[i]} in test group}

\item{\code{deltaRef}}{Posterior medians (first column), lower (second column) and upper (third column) 
limits of gene-specific biological cell-to-cell heterogeneity hyper-parameters \eqn{\delta[i]} in reference group}

\item{\code{phi}}{Posterior medians (first column), lower (second column) and upper (third column) 
limits of cell-specific mRNA content normalising constants \eqn{\phi[j]}}

\item{\code{s}}{Posterior medians (first column), lower (second column) and upper (third column) 
limits of cell-specific capture efficiency (or amplification biases if not using UMI based counts) normalising constants \eqn{s[j]}}

\item{\code{nu}}{Posterior medians (first column), lower (second column) and upper (third column) 
limits of cell-specific random effects \eqn{\nu[j]}}

\item{\code{thetaTest}}{Posterior median (first column), lower (second column) and upper (third column) 
limits of technical variability hyper-parameter \eqn{\theta[test]} for test sample (vector, all elements must be positive)}

\item{\code{thetaRef}}{Posterior median (first column), lower (second column) and upper (third column) 
limits of technical variability hyper-parameter \eqn{\theta[ref]} for reference sample (vector, all elements must be positive)}
}}
\examples{

# see help(BASiCS_DV_MCMC)

}

