% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FunctionsPart2.R
\name{BASiCS_DV_TestDE}
\alias{BASiCS_DV_TestDE}
\title{Detection of genes with changes in expression}
\usage{
BASiCS_DV_TestDE(Data, object, GeneNames, EpsilonM = 0.1, EpsilonD = 0.1,
  EviThresholdM = NULL, EviThresholdD = NULL,
  OrderVariable = "ProbDiffExp", GroupLabelRef = "Ref",
  GroupLabelTest = "Test", Plot = FALSE, OffSet = FALSE, EFDR_M = 0.05,
  EFDR_D = 0.05, ...)
}
\arguments{
\item{Data}{an object of class \code{\link[BASiCS]{BASiCS_DV_Data-class}}}

\item{object}{an object of class \code{\link[BASiCS]{BASiCS_DV_Chain-class}}}

\item{GeneNames}{to be removed. This will be an slot of Data}

\item{EpsilonM}{Minimum fold change tolerance threshold for detecting changes in overall expression (must be a positive real number)}

\item{EpsilonD}{Minimum fold change tolerance threshold for detecting changes in cell-to-cell biological over dispersion (must be a positive real number)}

\item{EviThresholdM}{Optional parameter. Evidence threshold for detecting changes in overall expression (must be a positive value, between 0 and 1)}

\item{EviThresholdD}{Optional parameter. Evidence threshold for detecting changes in cell-to-cell biological over dispersion (must be a positive value, between 0 and 1)}

\item{OrderVariable}{Ordering variable for output. Must take values in \code{c("GeneIndex", "GeneNames", "ProbDiffExp", "ProbDiffOverDisp")}.}

\item{GroupLabelRef}{Label assigned to reference group. Default: \code{GroupLabelRef = "Ref"}}

\item{GroupLabelTest}{Label assigned to reference group. Default: \code{GroupLabelRef = "Test"}}

\item{Plot}{If \code{Plot = T}, error rates control rates and volcano plots are generated.}

\item{...}{Graphical parameters (see \code{\link[graphics]{par}}).}
}
\value{
\code{BASiCS_DV_TestDE} returns a list of 4 elements:
\describe{
\item{\code{Table}}{A \code{\link[base]{data.frame}} containing the results of the differential expression test}
   \describe{
   \item{\code{Mu}}{Vector of length \code{q.bio}. For each biological gene, posterior median of gene-specific expression levels \eqn{\mu[i]}}
   \item{\code{Delta}}{Vector of length \code{q.bio}. For each biological gene, posterior median of gene-specific biological cell-to-cell heterogeneity hyper-parameter \eqn{\delta[i]}}
   \item{\code{Sigma}}{Vector of length \code{q.bio}. For each biological gene, proportion of the total variability that is due to a cell-to-cell biological heterogeneity component. }
   \item{\code{Prob}}{Vector of length \code{q.bio}. For each biological gene, probability of being highly variable according to the given thresholds.}
   \item{\code{HVG}}{Vector of length \code{q.bio}. For each biological gene, indicator of being detected as highly variable according to the given thresholds. }
   }
\item{\code{EviThreshold}}{Evidence thresholds.}
\item{\code{EFDR}}{Expected false discovery rate for the given thresholds.}
\item{\code{EFNR}}{Expected false negative rate for the given thresholds.}
}
}
\description{
Detection of genes with changes in expression
}
\details{
See vignette
}
\examples{
# See
help(BASiCS_DV_MCMC)
}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}
\seealso{
\code{\link[BASiCS]{BASiCS_DV_Chain-class}}
}

