% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASiCS_EffectiveSize.R
\name{BASiCS_EffectiveSize}
\alias{BASiCS_EffectiveSize}
\alias{BASiCS_effectiveSize}
\title{Calculate effective sample size for BASiCS_Chain parameters}
\usage{
BASiCS_EffectiveSize(object, Param, na.rm = TRUE)

BASiCS_effectiveSize(...)
}
\arguments{
\item{object}{an object of class \code{\linkS4class{BASiCS_Chain}}.}

\item{Param}{The parameter to use to calculate effectiveSize. Possible
values: \code{'mu'}, \code{'delta'}, \code{'phi'}, \code{'s'}, 
\code{'nu'}, \code{'theta'}, \code{'beta'}, \code{'sigma2'} and 
\code{'epsilon'}.}

\item{na.rm}{Remove \code{NA} values before calculating effectiveSize. 
Only relevant when \code{Param = "epsilon"} (genes with very low 
expression are excluding when infering the mean/over-dispersion trend. 
Default: \code{na.rm = TRUE}.}

\item{...}{Unused.}
}
\value{
A vector with effective sample sizes for all the elements 
of \code{Param}
}
\description{
A wrapper of \code{coda::effectiveSize} to be used with 
\code{\linkS4class{BASiCS_Chain}} objects.
}
\examples{

data(ChainSC)
BASiCS_EffectiveSize(ChainSC, Param = "mu")

}
