% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASiCS_LoadChain.R
\name{BASiCS_LoadChain}
\alias{BASiCS_LoadChain}
\title{Loads pre-computed MCMC chains generated by the \code{BASiCS_MCMC} function}
\usage{
BASiCS_LoadChain(RunName, StoreDir = getwd(), WithSpikes = TRUE)
}
\arguments{
\item{RunName}{String used to index `.Rds` files storing the MCMC chains (produced by the \code{BASiCS_MCMC} function)}

\item{StoreDir}{Directory where `.Rds` files is stored. Default: \code{StoreDir = getwd()}}

\item{WithSpikes}{\code{TRUE}/\code{FALSE} indicator of spike-ins usage. Default: \code{WithSpikes = TRUE}}
}
\value{
An object of class \code{\link[BASiCS]{BASiCS_Chain-class}}.
}
\description{
Loads pre-computed MCMC chains generated by the \code{BASiCS_MCMC} function, creating
a \code{\link[BASiCS]{BASiCS_Chain-class}} object
}
\examples{

Data = makeExampleBASiCS_Data()
Chain <- BASiCS_MCMC(Data, N = 50, Thin = 5, Burn = 5,
                           StoreChains = TRUE, StoreDir = tempdir(), 
                           RunName = "Test")
ChainLoad <- BASiCS_LoadChain(RunName = "Test", StoreDir = tempdir())

}
\references{
Vallejos, Marioni and Richardson (2015). Bayesian Analysis of Single-Cell Sequencing data.
}
\seealso{
\code{\link[BASiCS]{BASiCS_Chain-class}}
}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}
}
