% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BASiCS_MultiChain}
\alias{BASiCS_MultiChain}
\alias{BASiCS_MultiChain-class}
\title{The BASiCS_MultiChain class}
\description{
Container of several sets of MCMC samples of the BASiCS' model
parameters as generated by the functions \code{\link[BASiCS]{BASiCS_MCMC}}
or \code{\link[BASiCS]{BASiCS_MultiMCMC}}.
}
\section{Slots}{

\describe{
\item{\code{chains}}{List of \linkS4class{BASiCS_Chain} objects.}
}}

\examples{

# A BASiCS_Chain object created by the BASiCS_MCMC function.
Data <- makeExampleBASiCS_Data()

## run 4 chains
Chains <- BASiCS_MCMC(Data, N = 100, Thin = 2, Burn = 2, Regression = TRUE, NChains = 4)
## runs 4 chains by default
Chains <- BASiCS_MultiMCMC(Data, N = 100, Thin = 2, Burn = 2, Regression = TRUE)


}
