% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASiCS_PlotMultiChain.R
\name{BASiCS_PlotMultiChain}
\alias{BASiCS_PlotMultiChain}
\title{todo document
todo rethink x/y axes plus colour}
\usage{
BASiCS_PlotMultiChain(object, Param = "mu")
}
\description{
todo document
todo rethink x/y axes plus colour
}
