% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASiCS_DetectHVG_LVG.R
\name{BASiCS_PlotVG}
\alias{BASiCS_PlotVG}
\title{Plots of HVG/LVG search.}
\usage{
BASiCS_PlotVG(object, Plot = c("Grid", "VG"), ...)
}
\arguments{
\item{object}{\linkS4class{BASiCS_ResultVG} object.}

\item{Plot}{Character scalar specifying the type of plot to be made.
Options are "Grid" and "VG".}

\item{...}{Passed to \code{.VGPlot}.}
}
\value{
A plot.
}
\description{
Plots of HVG/LVG search.
}
\examples{
data(ChainSC)

# Highly and lowly variable genes detection (within a single group of cells)
DetectHVG <- BASiCS_DetectHVG(ChainSC, VarThreshold = 0.60,
                              EFDR = 0.10, Plot = TRUE)
BASiCS_PlotVG(DetectHVG)
}
