% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BASiCS_Result}
\alias{BASiCS_Result}
\alias{BASiCS_Result-class}
\title{The BASiCS_Result class}
\description{
Container of results for a single test (HVG/LVG/DE).
This should be an abstract class (but this is R so no)
and shouldn't be directly instantiated.
Defines a very small amount of common behaviour for
\linkS4class{BASiCS_ResultDE} and \linkS4class{BASiCS_ResultVG}.
}
\section{Slots}{

\describe{
\item{\code{Table}}{Tabular results for each gene.}

\item{\code{Name}}{The name of the test performed (typically "Mean", "Disp" or 
"ResDisp")}

\item{\code{ProbThreshold}}{Posterior probability threshold used in differential 
test.}

\item{\code{EFDR,EFNR}}{Expected false discovery and expected false negative rates
for differential test.}

\item{\code{Extra}}{Additional objects for class flexibility.}
}}

