% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BASiCS_ResultDE}
\alias{BASiCS_ResultDE}
\alias{BASiCS_ResultDE-class}
\title{The BASiCS_ResultDE class}
\description{
Container of results for a single differential test.
}
\section{Slots}{

\describe{
\item{\code{Table}}{Tabular results for each gene.}

\item{\code{Name}}{The name of the test performed (typically "Mean", "Disp" or 
"ResDisp")}

\item{\code{GroupLabel1,GroupLabel2}}{Group labels.}

\item{\code{ProbThreshold}}{Posterior probability threshold used in differential 
test.}

\item{\code{EFDR,EFNR}}{Expected false discovery and expected false negative rates
for differential test.}

\item{\code{EFDRgrid,EFNRgrid}}{Grid of EFDR and EFNR values check before thresholds
were fixed.}

\item{\code{Epsilon}}{Minimum fold change or difference threshold.}

\item{\code{Extra}}{objects for class flexibility.}
}}

