% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BASiCS_ResultsDE}
\alias{BASiCS_ResultsDE}
\alias{BASiCS_ResultsDE-class}
\title{The BASiCS_ResultsDE class}
\description{
Results of BASiCS_TestDE
}
\section{Slots}{

\describe{
\item{\code{Results}}{\code{\linkS4class{BASiCS_ResultDE}} objects}

\item{\code{Chain1,Chain2}}{\code{\linkS4class{BASiCS_Chain}} objects.}

\item{\code{GroupLabel1,GroupLabel2}}{Labels for Chain1 and Chain2}

\item{\code{Offset}}{Ratio between median of chains}

\item{\code{RowData}}{Annotation for genes}

\item{\code{Extras}}{Slot for extra information to be added later}
}}

