% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASiCS_ShowFit.R
\name{BASiCS_ShowMultiFit}
\alias{BASiCS_ShowMultiFit}
\title{todo document}
\usage{
BASiCS_ShowMultiFit(
  object,
  variance = 1.2,
  Uncertainty = TRUE,
  Variability = TRUE
)
}
\description{
todo document
}
