% Generated by roxygen2 (4.0.2): do not edit by hand
\name{displaySummaryMu}
\alias{displaySummaryDelta}
\alias{displaySummaryDelta,BASiCS_Summary-accessors}
\alias{displaySummaryMu}
\alias{displaySummaryMu,BASiCS_Summary-accessors}
\alias{displaySummaryNu}
\alias{displaySummaryNu,BASiCS_Summary-accessors}
\alias{displaySummaryPhi}
\alias{displaySummaryPhi,BASiCS_Summary-accessors}
\alias{displaySummaryS}
\alias{displaySummaryS,BASiCS_Summary-accessors}
\alias{displaySummaryTheta}
\alias{displaySummaryTheta,BASiCS_Summary-accessors}
\title{Accessors for the slots of a BASiCS_Summary object}
\usage{
displaySummaryMu(object)

displaySummaryDelta(object)

displaySummaryPhi(object)

displaySummaryS(object)

displaySummaryNu(object)

displaySummaryTheta(object)
}
\arguments{
\item{object}{an object of class \code{\link[BASiCS]{BASiCS_Summary-class}}}
}
\value{
\describe{
\item{\code{displaySummaryMu}}{The \code{mu} slot of an object of class \code{\link[BASiCS]{BASiCS_Summary-class}}.}
\item{\code{displaySummaryDelta}}{The \code{delta} slot of an object of class \code{\link[BASiCS]{BASiCS_Summary-class}}.}
\item{\code{displaySummaryPhi}}{The \code{phi} slot of an object of class \code{\link[BASiCS]{BASiCS_Summary-class}}.}
\item{\code{displaySummaryS}}{The \code{s} slot of an object of class \code{\link[BASiCS]{BASiCS_Summary-class}}.}
\item{\code{displaySummaryNu}}{The \code{nu} slot of an object of class \code{\link[BASiCS]{BASiCS_Summary-class}}.}
\item{\code{displaySummaryTheta}}{The \code{theta} slot of an object of class \code{\link[BASiCS]{BASiCS_Summary-class}}.}
}
}
\description{
Accessors for the slots of a \code{\link[BASiCS]{BASiCS_Summary-class}}
}
\examples{
# See
help(BASiCS_MCMC)
}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}
\references{
Vallejos, Marioni and Richardson (2015). Bayesian Analysis of Single-Cell Sequencing data.
}
\seealso{
\code{\link[BASiCS]{BASiCS_Summary-class}}
}

