% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{BASiCS_Summary-class}
\alias{BASiCS_Summary,BASiCS_Summary-class}
\alias{BASiCS_Summary-class}
\title{The BASiCS_Summary class}
\description{
This class contains valid values for BASiCS model parameters (see Vallejos et al, 2015)
}

