% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASiCS_TestDE.R
\name{BASiCS_TestDE}
\alias{BASiCS_TestDE}
\title{Detection of genes with changes in expression}
\usage{
BASiCS_TestDE(Chain1, Chain2, EpsilonM = 0.4 * log(2), EpsilonD = 0.4 *
  log(2), EviThresholdM = NULL, EviThresholdD = NULL,
  OrderVariable = "Prob", GroupLabel1 = "Group1", GroupLabel2 = "Group2",
  Plot = TRUE, PlotOffset = TRUE, OffSet = TRUE, EFDR_M = 0.05,
  EFDR_D = 0.05, GenesSelect = NULL, ...)
}
\arguments{
\item{Chain1}{an object of class \code{\link[BASiCS]{BASiCS_Chain-class}} containing parameter estimates for the first group of cells}

\item{Chain2}{an object of class \code{\link[BASiCS]{BASiCS_Chain-class}} containing parameter estimates for the second group of cells}

\item{EpsilonM}{Minimum fold change tolerance threshold for detecting changes in overall expression (must be a positive real number)}

\item{EpsilonD}{Minimum fold change tolerance threshold for detecting changes in cell-to-cell biological over dispersion (must be a positive real number)}

\item{EviThresholdM}{Optional parameter. Evidence threshold for detecting changes in overall expression (must be a positive value, between 0 and 1)}

\item{EviThresholdD}{Optional parameter. Evidence threshold for detecting changes in cell-to-cell biological over dispersion (must be a positive value, between 0 and 1)}

\item{OrderVariable}{Ordering variable for output. Must take values in \code{c("GeneIndex", "GeneNames", "Prob")}.}

\item{GroupLabel1}{Label assigned to reference group. Default: \code{GroupLabel1 = "Group1"}}

\item{GroupLabel2}{Label assigned to reference group. Default: \code{GroupLabel2 = "Group2"}}

\item{Plot}{If \code{Plot = TRUE}, MA and volcano plots are generated.}

\item{PlotOffset}{If \code{Plot = TRUE}, the offset effect is visualised.}

\item{OffSet}{Optional argument to remove a fix offset effect (if not previously removed from the MCMC chains). This argument will be removed shorly, once offset removal is built as an internal step.}

\item{EFDR_M}{Target for expected false discovery rate related to the comparison of means (default = 0.05)}

\item{EFDR_D}{Target for expected false discovery rate related to the comparison of dispersions (default = 0.05)}

\item{GenesSelect}{Optional argument to provide a user-defined list of genes to be considered for the comparison (default = NULL). When used, this argument must be a vector of 'TRUE' (include gene) / 'FALSE' (exclude gene) indicator, with the same length as the number of intrinsic genes and following the same order as how genes are displayed in the table of counts.  This argument is necessary in order to have a meaningful EFDR calibration when the user decides to exclude some genes from the comparison.}

\item{...}{Graphical parameters (see \code{\link[graphics]{par}}).}
}
\value{
\code{BASiCS_TestDE} returns a list of 4 elements:
\describe{
\item{\code{TableMean}}{A \code{\link[base]{data.frame}} containing the results of the differential mean test}
   \describe{
   \item{\code{GeneNames}}{Gene name}
   \item{\code{MeanOverall}}{For each gene, the estimated mean expression parameter \eqn{\mu[i]} is averaged across both groups of cells (weighted by sample size).}
   \item{\code{Mean1}}{Estimated mean expression parameter \eqn{\mu[i]} for each biological gene in the first group of cells.}
   \item{\code{Mean2}}{Estimated mean expression parameter \eqn{\mu[i]} for each biological gene in the second group of cells.}
   \item{\code{MeanFC}}{Fold change in mean expression parameters between the first and second groups of cells.}
   \item{\code{MeanLog2FC}}{Log2-transformed fold change in mean expression between the first and second groups of cells.}
   \item{\code{ProbDiffMean}}{Posterior probability for mean expression difference between the first and second groups of cells.}
   \item{\code{ResultDiffExp}}{Indicator if a gene has a higher mean expression in the first or second groups of cells.}
   }
\item{\code{TableDisp}}{A \code{\link[base]{data.frame}} containing the results of the differential dispersion test (excludes genes for which the mean changes).}
   \describe{
   \item{\code{GeneNames}}{Gene name}
   \item{\code{DispOverall}}{For each gene, the estimated over-dispersion parameter \eqn{\delta[i]} is averaged across both groups of cells (weighted by sample size).}
   \item{\code{Disp1}}{Estimated over-dispersion parameter \eqn{\delta[i]} for each biological gene in the first group of cells.}
   \item{\code{Disp2}}{Estimated over-dispersion parameter \eqn{\delta[i]} for each biological gene in the second group of cells.}
   \item{\code{DispFC}}{Fold change in over-dispersion parameters between the between the first and second groups of cells.}
   \item{\code{DispLog2FC}}{Log-transformed fold change in over-dispersion between the first and second groups of cells.}
   \item{\code{ProbDiffDisp}}{Posterior probability for over-dispersion difference between the first and second groups of cells.}
   \item{\code{ResultDiffDisp}}{Indicator if a gene has a higher over-dispersion in the first or second groups of cells.}
   }
 \item{\code{DiffExpSummary}}{A list containing the following information for the differential mean expression test:}
   \describe{
  \item{\code{EviThreshold}}{Evidence thresholds.}
  \item{\code{EFDR}}{Expected false discovery rate for the given thresholds.}
  \item{\code{EFNR}}{Expected false negative rate for the given thresholds.}
  }
 \item{\code{DiffOverDispSummary}}{A list containing the following information for the differential over-dispersion test:}
   \describe{
  \item{\code{EviThreshold}}{Evidence thresholds.}
  \item{\code{EFDR}}{Expected false discovery rate for the given thresholds.}
  \item{\code{EFNR}}{Expected false negative rate for the given thresholds.}
  }
 \item{\code{Chain1_offset}}{an \code{\link[BASiCS]{BASiCS_Chain-class}} object: \code{Chain1} after offset removal.}
 \item{\code{Chain2_offset}}{an \code{\link[BASiCS]{BASiCS_Chain-class}} object: \code{Chain2} after offset removal
 (this is only provided for completeness; \code{Chain2} is not affected by the offset).}
 \item{\code{OffsetChain}}{MCMC chain calculated for the offset effect.}
 \item{\code{Offset}}{Estimated offset (posterior median of \code{OffsetChain}). Default value set equal to 1 when offset correction is not performed.}
}
}
\description{
Function to assess changes in expression (mean and over-dispersion)
}
\examples{

# Loadind two synthetic datasets
Data1 <- makeExampleBASiCS_Data(Example = 1)
Data2 <- makeExampleBASiCS_Data(Example = 2)

# Running the MCMC (separately for each group of cells) 
Chain1 <- BASiCS_MCMC(Data = Data1, N = 20000, Thin = 20, Burn = 10000, 
                      PrintProgress = FALSE)
Chain2 <- BASiCS_MCMC(Data = Data2, N = 20000, Thin = 20, Burn = 10000, 
                      PrintProgress = FALSE)
                      
Test <- BASiCS_TestDE(Chain1 = Chain1, Chain2 = Chain2,
                      EpsilonM = 0.4*log(2), EpsilonD = 0.4*log(2), OffSet = TRUE)

}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl} and Nils Eling
}
