% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FunctionsPart1.R
\name{BASiCS_VarianceDecomp}
\alias{BASiCS_VarianceDecomp}
\title{Decomposition of gene expression variability according to BASiCS}
\usage{
BASiCS_VarianceDecomp(Data, object, OrderVariable = "BioVarGlobal",
  Plot = TRUE, ...)
}
\arguments{
\item{Data}{an object of class \code{\link[BASiCS]{BASiCS_Data-class}}}

\item{object}{an object of class \code{\link[BASiCS]{BASiCS_Chain-class}}}

\item{OrderVariable}{Ordering variable for output. Must take values in \code{c("GeneNames", "BioVarGlobal", "TechVarGlobal", "ShotNoiseGlobal")}.}

\item{Plot}{If \code{TRUE}, a barplot of the variance decomposition (global and by batches, if any) is generated}

\item{...}{Other arguments to be passed to \code{\link[graphics]{barplot}}}

\item{GeneNames}{Vector of characters containing gene names (must be in the same order as in the Counts matrix).}
}
\value{
A \code{\link[base]{data.frame}} whose first 4 columns correspond to
\describe{
\item{\code{GeneName}}{Gene name (as indicated by user)}
\item{\code{BioVarGlobal}}{Percentage of variance explained by a biological cell-to-cell heterogeneity component (overall across all cells)}
\item{\code{TechVarGlobal}}{Percentage of variance explained by the technical cell-to-cell heterogeneity component (overall across all cells)}
\item{\code{ShotNoiseGlobal}}{Percentage of variance explained by the shot noise component (baseline, overall across all cells)}
}
If more than 1 batch of cells are being analysed, the remaining columns contain the corresponding variance decomposition calculated within each batch.
}
\description{
Decomposition of gene expression variability according to BASiCS
}
\details{
See vignette
}
\examples{
# See
help(BASiCS_MCMC)
}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}
\references{
Vallejos, Marioni and Richardson (2015). Bayesian Analysis of Single-Cell Sequencing data.
}
\seealso{
\code{\link[BASiCS]{BASiCS_Chain-class}}
}

