% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffers.R
\name{buffers}
\alias{buffers}
\alias{buffer300,}
\alias{buffer20}
\alias{Lines}
\alias{parallel}
\alias{to}
\alias{the}
\alias{coast}
\alias{at}
\alias{different}
\alias{distances}
\alias{(in}
\alias{km)}
\title{Coastal parallel lines}
\description{
Coastal parallel lines
}
\details{
See the vignette on how the data were produced for the code: \code{vignette("10 - Save Data", package = "basics")}
}
\keyword{data}
