% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\docType{methods}
\name{colnames}
\alias{colnames}
\alias{colnames,BASiCS_Chain-method}
\title{'colnames' method for BASiCS_Chain objects}
\usage{
\S4method{colnames}{BASiCS_Chain}(x)
}
\arguments{
\item{x}{A \code{\linkS4class{BASiCS_Chain}} object.}
}
\value{
An vector of labels
}
\description{
Returns the labels of cell-specific BASiCS parameters
}
\examples{

# see help(BASiCS_MCMC)

}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}
}
