% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsPart2.R
\name{CombineBASiCS_Chain}
\alias{CombineBASiCS_Chain}
\title{Creates a BASiCS_D_Chain object based on 2 independent `BASiCS_Chain` objects}
\usage{
CombineBASiCS_Chain(ChainTest, ChainRef)
}
\arguments{
\item{ChainTest}{A \code{\link[BASiCS]{BASiCS_Chain-class}} object (test group)}

\item{ChainRef}{A \code{\link[BASiCS]{BASiCS_Chain-class}} object (reference group)}
}
\value{
An object of class \code{\link[BASiCS]{BASiCS_D_Chain-class}}.
}
\description{
Creates a \code{\link[BASiCS]{BASiCS_D_Chain-class}} object from 
based on 2 independent  \code{\link[BASiCS]{BASiCS_Chain-class}} objects
}
\examples{

# See vignette

}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}
}
\seealso{
\code{\link[BASiCS]{BASiCS_D_Data-class}}
}

