% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsPart2.R
\name{CombineBASiCS_Data}
\alias{CombineBASiCS_Data}
\title{Creates a BASiCS_D_Data object based on 2 independent `BASiCS_Data` objects}
\usage{
CombineBASiCS_Data(DataTest, DataRef)
}
\arguments{
\item{DataTest}{A \code{\link[BASiCS]{BASiCS_Data-class}} object (test group)}

\item{DataRef}{A \code{\link[BASiCS]{BASiCS_Data-class}} object (reference group)}
}
\value{
An object of class \code{\link[BASiCS]{BASiCS_D_Data-class}}.
}
\description{
Creates a \code{\link[BASiCS]{BASiCS_D_Data-class}} object from 
based on 2 independent  \code{\link[BASiCS]{BASiCS_Data-class}} objects
}
\examples{

DataTest = makeExampleBASiCS_Data()
DataRef = makeExampleBASiCS_Data()
Data = CombineBASiCS_Data(DataTest, DataRef)

}
\seealso{
\code{\link[BASiCS]{BASiCS_D_Data-class}}
}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}
}
