% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FunctionsPart1.R
\name{DenoisedRates}
\alias{BASiCS_DenoisedRates}
\alias{DenoisedRates}
\title{Calculates normalised and denoised expression rates}
\usage{
BASiCS_DenoisedRates(Data, Chain, PrintProgress = F, Propensities = T)
}
\arguments{
\item{Data}{an object of class \code{\link[BASiCS]{BASiCS_Data-class}}}

\item{Chain}{an object of class \code{\link[BASiCS]{BASiCS_Chain-class}}}

\item{PrintProgress}{If \code{T}, partial progress information is printed in the console.}

\item{Propensities}{If \code{T}, returns underlying expression propensitites \eqn{\rho[ij]}. Otherwise, denoised rates \eqn{\mu[i] \rho[ij]} are returned.}
}
\value{
A matrix of normalised and denoised expression counts.
}
\description{
Calculates normalised and denoised expression rates, by removing the effect of technical variation.
}
\details{
See vignette
}
\examples{
# See
help(BASiCS_MCMC)
}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}
\seealso{
\code{\link[BASiCS]{BASiCS_Data-class}} \code{\link[BASiCS]{BASiCS_Chain-class}}
}

