% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\docType{methods}
\name{dim}
\alias{dim}
\alias{dim,BASiCS_Chain-method}
\title{'dim' method for BASiCS_Chain objects}
\usage{
\S4method{dim}{BASiCS_Chain}(x)
}
\arguments{
\item{x}{A \code{\linkS4class{BASiCS_Chain}} object.}
}
\value{
An vector of dimensions
}
\description{
Returns the dimensions (genes x cells) of a 
 \code{\linkS4class{BASiCS_Chain}}
}
\examples{

data(ChainSC)
dimnames(ChainSC)

}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}
}
