% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\docType{methods}
\name{dimnames}
\alias{dimnames}
\alias{dimnames,BASiCS_Chain-method}
\title{'dimnames' method for BASiCS_Chain objects}
\usage{
\S4method{dimnames}{BASiCS_Chain}(x)
}
\arguments{
\item{x}{A \code{\linkS4class{BASiCS_Chain}} object.}
}
\value{
A list of two elements: (1) a vector of gene names and
(2) a vector of cell names.
}
\description{
Returns the dimension names (genes x cells) of a 
 \code{\linkS4class{BASiCS_Chain}}
}
\examples{

data(ChainSC)
dimnames(ChainSC)

}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}
}
