% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Methods.R
\docType{methods}
\name{displayChainBASiCS-BASiCS_DV_Chain-method}
\alias{displayChainBASiCS}
\alias{displayChainBASiCS,BASiCS_DV_Chain-method}
\alias{displayChainBASiCS-BASiCS_DV_Chain-method}
\title{Accessors for the slots of a BASiCS_DV_Chain object}
\usage{
\S4method{displayChainBASiCS}{BASiCS_DV_Chain}(object, Param = "muTest")
}
\arguments{
\item{object}{an object of class \code{\link[BASiCSDV]{BASiCS_DV_Chain-class}}}

\item{Param}{Name of the slot to be used for the accessed.
Possible values: \code{muTest, muRef, deltaTest, deltaRef, phi, s, nu, thetaTest, thetaRef}}
}
\value{
The requested slot of an object of class \code{\link[BASiCSDV]{BASiCS_DV_Chain-class}}
}
\description{
Accessors for the slots of a \code{\link[BASiCSDV]{BASiCS_DV_Chain-class}}
}
\examples{
# See
help(BASiCS_DV_MCMC)
}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}
\seealso{
\code{\link[BASiCSDV]{BASiCS_DV_Chain-class}}
}

