% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\docType{methods}
\name{displaySummaryBASiCS-BASiCS_D_Summary-method}
\alias{displaySummaryBASiCS}
\alias{displaySummaryBASiCS,BASiCS_D_Summary-method}
\alias{displaySummaryBASiCS-BASiCS_D_Summary-method}
\title{Accessors for the slots of a BASiCS_D_Summary object}
\usage{
\S4method{displaySummaryBASiCS}{BASiCS_D_Summary}(object, Param = "muTest")
}
\arguments{
\item{object}{an object of class \code{\link[BASiCS]{BASiCS_D_Summary-class}}}

\item{Param}{Name of the slot to be used for the accessed. 
Possible values: \code{"muTest", "muRef", "deltaTest", "deltaRef", "phi", "s", "nu", "thetaTest", "thetaRef"}}
}
\value{
The requested slot of an object of class \code{\link[BASiCS]{BASiCS_D_Summary-class}}
}
\description{
Accessors for the slots of a \code{\link[BASiCS]{BASiCS_D_Summary-class}}
}
\examples{

# See
help(BASiCS_D_MCMC)
  
}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}
\seealso{
\code{\link[BASiCS]{BASiCS_D_Summary-class}}
}

