% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{format,BASiCS_ResultsDE-method}
\alias{format,BASiCS_ResultsDE-method}
\alias{format,BASiCS_ResultDE-method}
\alias{format,BASiCS_ResultVG-method}
\title{Methods for formatting \linkS4class{BASiCS_Result} and 
\linkS4class{BASiCS_ResultsDE} objects.}
\usage{
\S4method{format}{BASiCS_ResultsDE}(x, Parameter, Filter = TRUE, ProbThreshold = NULL, ...)

\S4method{format}{BASiCS_ResultDE}(x, Filter = TRUE, ProbThreshold = NULL, ...)

\S4method{format}{BASiCS_ResultVG}(x, Filter = TRUE, ProbThreshold = NULL, ...)
}
\arguments{
\item{x}{Object being subsetted.}

\item{Parameter}{Character scalar indicating which of the 
\linkS4class{BASiCS_Result} should be formatted.}

\item{Filter}{Logical scalar indicating whether results should be
filtered based on differential expression or HVG/LVG status if
\code{ProbThreshold=NULL}, or a probability threshold if
\code{ProbThreshold=NULL}}

\item{ProbThreshold}{Probability threshold to be used if \code{Filter=TRUE}}

\item{...}{Passed to \code{\link{format}}.}
}
\value{
A \code{data.frame}.
}
\description{
Methods for formatting \linkS4class{BASiCS_Result} and 
\linkS4class{BASiCS_ResultsDE} objects.
}
