% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsPart2.R
\name{makeExampleBASiCS_D_Data}
\alias{makeExampleBASiCS_D_Data}
\title{Create a simple example of a BASiCS_D_Data object with random data}
\usage{
makeExampleBASiCS_D_Data()
}
\value{
An object of class \code{\link[BASiCS]{BASiCS_D_Data-class}}, simulated from the model implemented in BASiCS (for 2 populations of cells). 
It contains 120 genes (100 biological and 20 spike-in) and 20 cells (10 in each sample).
}
\description{
A simple \code{\link[BASiCS]{BASiCS_D_Data-class}} object is generated by simulating a dataset from the
model in BASiCS. This is used for the examples in the package and the vignette.
}
\examples{
Data = makeExampleBASiCS_D_Data()
is(Data, "BASiCS_D_Data")

}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}
}
