% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FunctionsPart2.R
\name{makeExampleBASiCS_DV_Data}
\alias{makeExampleBASiCS_DV_Data}
\title{Create a simple example of a BASiCS_DV_Data object with random data}
\usage{
makeExampleBASiCS_DV_Data()
}
\value{
An object of class \code{\link[BASiCS]{BASiCS_DV_Data-class}}, simulated from the model implemented in BASiCSDV.
It contains 120 genes (100 biological and 20 spike-in) and 20 cells (10 in each sample).
}
\description{
A simple \code{\link[BASiCS]{BASiCS_DV_Data-class}} object is generated by simulating a dataset from the
model in BASiCSDV. This is used for the examples in the package and the vignette.
}
\examples{
Data = makeExampleBASiCS_DV_Data()
is(Data, "BASiCS_DV_Data")
}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}

