% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASiCS_DivideAndConquer.R
\name{multi_MCMC}
\alias{multi_MCMC}
\title{Run divide and conquer MCMC with BASiCS}
\usage{
multi_MCMC(
  Data,
  NSubsets = 5,
  SubsetBy = c("cell", "gene"),
  Alpha = 0.05,
  WithSpikes = FALSE,
  Regression = TRUE,
  mc.cores = min(detectCores(all.tests = FALSE, logical = TRUE), NSubsets),
  Seed = 42,
  Prior,
  MinGenesPerRBF = 100,
  ...
)
}
\arguments{
\item{Data}{SingleCellExperiemnt object}

\item{NSubsets}{Number of partitions to make}

\item{SubsetBy}{Passed to generateSubsets}

\item{Alpha}{Passed to generateSubsets}

\item{WithSpikes}{Similar to the argument in BASiCS_MCMC}

\item{mc.cores}{Number of cores for parallel processing}

\item{...}{Passed to BASiCS_MCMC}
}
\description{
Run divide and conquer MCMC with BASiCS
}
