% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newBASiCS_Chain.R
\name{newBASiCS_Chain}
\alias{newBASiCS_Chain}
\title{Creates a BASiCS_Chain object from pre-computed MCMC chains}
\usage{
newBASiCS_Chain(mu, delta, phi, s, nu, theta)
}
\arguments{
\item{mu}{MCMC chain for gene-specific expression levels \eqn{\mu_i}, defined as true input molecules in case of technical genes
(matrix with \code{q} columns, technical genes located at the end of the matrix, all elements must be positive numbers)}

\item{delta}{MCMC chain for gene-specific biological over-dispersion parameters \eqn{\delta_i}, biological genes only
(matrix with \code{q.bio} columns, all elements must be positive numbers)}

\item{phi}{MCMC chain for cell-specific mRNA content normalising parameters \eqn{\phi_j}
(matrix with \code{n} columns, all elements must be positive numbers and the sum of its elements must be equal to \code{n})}

\item{s}{MCMC chain for cell-specific technical normalisation parameters \eqn{s_j}
(matrix with \code{n} columns, all elements must be positive numbers)}

\item{nu}{MCMC chain for cell-specific random effects \eqn{\nu_j}
(matrix with \code{n} columns, all elements must be positive numbers)}

\item{theta}{MCMC chain for technical variability hyper-parameter \eqn{\theta} (vector, all elements must be positive)}
}
\value{
An object of class \code{\link[BASiCS]{BASiCS_Chain}}.
}
\description{
\code{BASiCS_Chain} creates a \code{\link[BASiCS]{BASiCS_Chain}} object from pre-computed MCMC chains.
}
\examples{

Data = makeExampleBASiCS_Data()
Chain <- BASiCS_MCMC(Data, N = 50, Thin = 5, Burn = 5)

ChainMu = displayChainBASiCS(Chain, "mu")
ChainDelta = displayChainBASiCS(Chain, "delta")
ChainPhi = displayChainBASiCS(Chain, "phi")
ChainS = displayChainBASiCS(Chain, "s")
ChainNu = displayChainBASiCS(Chain, "nu")
ChainTheta = displayChainBASiCS(Chain, "theta")

ChainNew <- newBASiCS_Chain(mu = ChainMu, delta = ChainDelta,
                            phi = ChainPhi, s = ChainS, 
                            nu = ChainNu, theta = ChainTheta)

}
\references{
Vallejos, Marioni and Richardson (2015). Bayesian Analysis of Single-Cell Sequencing data. PLoS Computational Biology.
}
\seealso{
\code{\link[BASiCS]{BASiCS_Chain}}
}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}
}
