% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsPart2.R
\name{newBASiCS_D_Data}
\alias{newBASiCS_D_Data}
\title{Creates a BASiCS_D_Data object from expression counts matrices and experimental information about spike-in genes}
\usage{
newBASiCS_D_Data(CountsTest, CountsRef, Tech, SpikeInputTest, SpikeInputRef,
  BatchInfoTest = NULL, BatchInfoRef = NULL, GeneNames)
}
\arguments{
\item{CountsTest}{Matrix of dimensions \code{q} times \code{n_test} whose elements corresponds to the raw expression counts in the test sample. 
First \code{q.bio} rows correspond to biological genes. Last \code{q-q.bio} rows correspond to technical spike-in genes.}

\item{CountsRef}{Matrix of dimensions \code{q} times \code{n_ref} whose elements corresponds to the raw expression counts in the reference sample. 
First \code{q.bio} rows correspond to biological genes. Last \code{q-q.bio} rows correspond to technical spike-in genes.}

\item{Tech}{Logical vector of length \code{q}. If \code{Tech = F} the gene is biological; otherwise the gene is spike-in.}

\item{SpikeInputTest}{Vector of length \code{q-q.bio} whose elements indicate the input number of molecules for the spike-in genes in the test sample (amount per cell).}

\item{SpikeInputRef}{Vector of length \code{q-q.bio} whose elements indicate the input number of molecules for the spike-in genes in the reference sample (amount per cell).}

\item{BatchInfoTest}{Vector of length \code{n_test} indicating experimental batches for test group.}

\item{BatchInfoRef}{Vector of length \code{n_ref} indicating experimental batches for test group.}

\item{GeneNames}{Vector of gene ids.}
}
\value{
An object of class \code{\link[BASiCS]{BASiCS_D_Data-class}}.
}
\description{
\code{newBASiCS_D_Data} creates a \code{\link[BASiCS]{BASiCS_D_Data-class}} object from 
expression counts matrices and experimental information about spike-in genes.
}
\examples{

# Rather than using this function, we recomend to use 'CombineBASiCS_Data'

}
\seealso{
\code{\link[BASiCS]{BASiCS_D_Data-class}}
}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}
}
