% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FunctionsPart2.R
\name{newBASiCS_DV_Data}
\alias{newBASiCS_DV_Data}
\title{Creates a BASiCS_DV_Data object from expression counts matrices and experimental information about spike-in genes}
\usage{
newBASiCS_DV_Data(CountsTest, CountsRef, Tech, SpikeInputTest, SpikeInputRef)
}
\arguments{
\item{CountsTest}{Matrix of dimensions \code{q} times \code{n_test} whose elements corresponds to the raw expression counts in the test sample.
First \code{q.bio} rows correspond to biological genes. Last \code{q-q.bio} rows correspond to technical spike-in genes.}

\item{CountsRef}{Matrix of dimensions \code{q} times \code{n_ref} whose elements corresponds to the raw expression counts in the reference sample.
First \code{q.bio} rows correspond to biological genes. Last \code{q-q.bio} rows correspond to technical spike-in genes.}

\item{Tech}{Logical vector of length \code{q}. If \code{Tech = F} the gene is biological; otherwise the gene is spike-in.}

\item{SpikeInputTest}{Vector of length \code{q-q.bio} whose elements indicate the input number of molecules for the spike-in genes in the test sample (amount per cell).}

\item{SpikeInputRef}{Vector of length \code{q-q.bio} whose elements indicate the input number of molecules for the spike-in genes in the reference sample (amount per cell).}
}
\value{
An object of class \code{\link[BASiCS]{BASiCS_DV_Data-class}}.
}
\description{
\code{newBASiCS_DV_Data} creates a \code{\link[BASiCS]{BASiCS_DV_Data-class}} object from
expression counts matrices and experimental information about spike-in genes.
}
\examples{
CountsTest = matrix(rpois(100*5, 2), ncol = 5)
CountsRef = matrix(rpois(100*5, 2), ncol = 5)
Tech = c(rep(FALSE,80),rep(TRUE,20))
SpikeInputTest = rgamma(20,1,1)
SpikeInputRef = rgamma(20,1,1)
Data = newBASiCS_DV_Data(CountsTest, CountsRef, Tech, SpikeInputTest, SpikeInputRef)
}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}
\seealso{
\code{\link[BASiCS]{BASiCS_DV_Data-class}}
}

