% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\docType{methods}
\name{plot-BASiCS_D_Chain-method}
\alias{plot}
\alias{plot,BASiCS_D_Chain-method}
\alias{plot-BASiCS_D_Chain-method}
\title{'plot' method for BASiCS_D_Chain objects}
\usage{
\S4method{plot}{BASiCS_D_Chain,ANY}(x, Param = "muTest", Gene = NULL,
  Cell = NULL, ylab = "", xlab = "", main = "", ...)
}
\arguments{
\item{x}{A \code{BASiCS_D_Chain} object.}

\item{Param}{Name of the slot to be used for the plot. Possible values: \code{muTest, muRef, deltaTest, deltaRef, phi, s, nu, thetaTest, thetaRef}}

\item{Gene}{Specifies which gene is requested. Required only if \code{Param = "muTest"}, \code{Param = "muRef"}, \code{"deltaTest"} or \code{"deltaRef"}}

\item{Cell}{Specifies which cell is requested. Required only if \code{Param = "phi", "s"} or \code{"nu"}}

\item{ylab}{As in \code{\link[graphics]{par}}.}

\item{xlab}{As in \code{\link[graphics]{par}}.}

\item{main}{As in \code{\link[graphics]{par}}.}

\item{...}{Other graphical parameters (see \code{\link[graphics]{par}}).}
}
\description{
'plot' method for BASiCS_D_Chain objects
}
\examples{

# See
help(BASiCS_D_MCMC)

}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}

