% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\docType{methods}
\name{plot-BASiCS_D_Summary-method}
\alias{plot-BASiCS_D_Summary-method}
\alias{plot,BASiCS_D_Summary-method}
\title{'plot' method for BASiCS_D_Summary objects}
\usage{
\S4method{plot}{BASiCS_D_Summary,ANY}(x, Param = "muTest", Param2 = NULL,
  Genes = NULL, Cells = NULL, xlab = "", ylab = "", xlim = "",
  ylim = "", main = "", pch = 16, col = "blue", bty = "n", ...)
}
\arguments{
\item{x}{A \code{BASiCS_D_Summary} object.}

\item{Param}{Name of the slot to be used for the plot. 
Possible values: \code{"muTest", "muRef", "deltaTest", "deltaRef", "phi", "s", "nu", "thetaTest", "thetaRef"}}

\item{Param2}{Name of the second slot to be used for the plot. 
Possible values: \code{"muTest", "muRef", "deltaTest", "deltaRef", "phi", "s", "nu"} 
(combinations between gene-specific and cell-specific parameters are not allowed)}

\item{Genes}{Specifies which genes are requested. 
Required only if \code{Param = "muTest", "muRef", "deltaTest"} or \code{"deltaRef"}}

\item{Cells}{Specifies which cells are requested. 
Required only if \code{Param = "phi", "s"} or \code{"nu"}}

\item{xlab}{As in \code{\link[graphics]{par}}.}

\item{ylab}{As in \code{\link[graphics]{par}}.}

\item{xlim}{As in \code{\link[graphics]{par}}.}

\item{ylim}{As in \code{\link[graphics]{par}}.}

\item{main}{As in \code{\link[graphics]{par}}.}

\item{pch}{As in \code{\link[graphics]{par}}.}

\item{col}{As in \code{\link[graphics]{par}}.}

\item{bty}{As in \code{\link[graphics]{par}}.}

\item{...}{Other graphical parameters (see \code{\link[graphics]{par}}).}
}
\description{
'plot' method for BASiCS_D_Summary objects
}
\examples{

# See
help(BASiCS_D_MCMC)

}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}
