% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,BASiCS_Chain-method}
\title{'plot' method for BASiCS_Chain and BASiCS_Summary objects}
\usage{
\S4method{plot}{BASiCS_Chain}(x, y, Param = "mu", Column = 1, ylab = "",
  xlab = "", bty = "", ...)
}
\description{
'plot' method for BASiCS_Chain and BASiCS_Summary objects
}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}
\references{
Vallejos, Marioni and Richardson (2015). Bayesian Analysis of Single-Cell Sequencing data.
}

