% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\docType{methods}
\name{Plots of diagnostic measure for MCMC parameters}
\alias{Plots of diagnostic measure for MCMC parameters}
\alias{diagPlot}
\alias{diagPlot-method}
\title{Create diagnostic plots of MCMC parameters}
\usage{
diagPlot(object, ...)
}
\arguments{
\item{object}{an object of class \code{\linkS4class{BASiCS_Summary}}}

\item{Param}{Optional name of a chain parameter to restrict the histogram;
if not supplied, all parameters will be assessed.
Possible values: \code{'mu'}, \code{'delta'}, \code{'phi'},
\code{'s'}, \code{'nu'}, \code{'theta'}, \code{'beta'},
\code{'sigma2'} and \code{'epsilon'}.}

\item{Measure}{The diagnostic measure to be plotted in the histogram.
Possible values: \code{"effectiveSize"}, \code{"geweke.diag"}.}

\item{x, y}{MCMC parameter values to be plotted on the x or y axis, 
respectively.}

\item{LogX, LogY}{A boolean value indicating whether to use a log10
transformation for the x or y axis, respectively.}
}
\value{
A ggplot object.
}
\description{
Plot parameter values and effective sample size or 
Geweke's diagnostic z-statistic. See \link[coda]{effectiveSize} and 
\link[coda]{geweke.diag} for more details on these diagnostic measures.
}
\examples{

# See
help(BASiCS_MCMC)

}
\seealso{
\code{\linkS4class{BASiCS_Chain}}
}
\author{
Alan O'Callaghan \email{a.b.ocallaghan@sms.ed.ac.uk}
}
