% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\docType{methods}
\name{rownames}
\alias{rownames}
\alias{rownames,BASiCS_Chain-method}
\title{'rownames' method for BASiCS_Chain objects}
\usage{
\S4method{rownames}{BASiCS_Chain}(x)
}
\arguments{
\item{x}{A \code{\linkS4class{BASiCS_Chain}} object.}
}
\value{
An vector of labels
}
\description{
Returns the labels of gene-specific BASiCS parameters
}
\examples{

data(ChainSC)
rownames(ChainSC)

}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}
}
