% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\docType{methods}
\name{show,BASiCS_OffsetCorrected-method}
\alias{show,BASiCS_OffsetCorrected-method}
\title{Accessors for the slots of a \code{\linkS4class{BASiCS_Summary}} object}
\usage{
\S4method{show}{BASiCS_OffsetCorrected}(object)
}
\arguments{
\item{object}{an object of class \code{\linkS4class{BASiCS_OffsetCorrected}}}
}
\description{
Accessors for the slots of a \code{\linkS4class{BASiCS_Summary}} object
}
\examples{

help(BASiCS_MCMC)

}
\seealso{
\code{\link[methods]{show}}
}
