% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\docType{methods}
\name{show,BASiCS_ResultDE-method}
\alias{show,BASiCS_ResultDE-method}
\title{Accessors for the slots of a \code{\linkS4class{BASiCS_ResultDE}} object}
\usage{
\S4method{show}{BASiCS_ResultDE}(object)
}
\arguments{
\item{object}{an object of class \code{\linkS4class{BASiCS_ResultDE}}}
}
\description{
Accessors for the slots of a \code{\linkS4class{BASiCS_ResultDE}} object
}
\examples{

help(BASiCS_MCMC)

}
\seealso{
\code{\link[methods]{show}}
}
