% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\docType{methods}
\name{show,BASiCS_ResultsDE-method}
\alias{show,BASiCS_ResultsDE-method}
\title{Accessors for the slots of a \code{\linkS4class{BASiCS_ResultsDE}}
object}
\usage{
\S4method{show}{BASiCS_ResultsDE}(object)
}
\arguments{
\item{object}{an object of class \code{\linkS4class{BASiCS_ResultsDE}}}
}
\value{
Prints a summary of the properties of \code{object}.
}
\description{
Accessors for the slots of a \code{\linkS4class{BASiCS_ResultsDE}}
object
}
\examples{

help(BASiCS_MCMC)

}
\seealso{
\code{\link[methods]{show}}
}
