% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{[,BASiCS_ResultsDE,ANY,ANY,ANY-method}
\alias{[,BASiCS_ResultsDE,ANY,ANY,ANY-method}
\alias{[[,BASiCS_ResultsDE,ANY,ANY-method}
\alias{[,BASiCS_Result,ANY,ANY,ANY-method}
\title{Methods for subsetting \linkS4class{BASiCS_Result} and 
\linkS4class{BASiCS_ResultsDE} objects.}
\usage{
\S4method{[}{BASiCS_ResultsDE,ANY,ANY,ANY}(x, i, j, drop = FALSE)

\S4method{[[}{BASiCS_ResultsDE,ANY,ANY}(x, i)

\S4method{[}{BASiCS_Result,ANY,ANY,ANY}(x, i, j, drop = FALSE)
}
\arguments{
\item{x}{Object being subsetted.}

\item{i}{See \code{?`[`}, \code{?`[[`}}

\item{j, drop}{Ignored.}
}
\value{
An object of the same class as \code{x}.
}
\description{
Methods for subsetting \linkS4class{BASiCS_Result} and 
\linkS4class{BASiCS_ResultsDE} objects.
}
