% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\docType{methods}
\name{Summary}
\alias{Summary}
\alias{Summary,BASiCS_D_Chain-method}
\title{'Summary' method for BASiCS_D_Chain objects}
\usage{
\S4method{Summary}{BASiCS_D_Chain}(x, prob = 0.95)
}
\arguments{
\item{x}{A \code{BASiCS_D_Chain} object.}

\item{prob}{\code{prob} argument for \code{\link[coda]{HPDinterval}} function.}
}
\value{
An object of class \code{\link[BASiCS]{BASiCS_D_Summary-class}}.
}
\description{
For each of the BASiCS parameters, 
\code{Summary} returns the corresponding postior medians and limits of the high posterior
density interval with probabilty equal to \code{prob}.
}
\examples{

# See help(BASiCS_D_MCMC)

}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}
