% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{Summary}
\alias{Summary}
\alias{Summary,BASiCS_Chain-method}
\title{'Summary' method for BASiCS_Chain objects}
\usage{
\S4method{Summary}{BASiCS_Chain}(x, prob = 0.95, ..., na.rm = FALSE)
}
\arguments{
\item{x}{object A \code{BASiCS_Chain} object.}

\item{prob}{\code{prob} argument for function \code{\link[coda]{HPDinterval}}}
}
\value{
An object of class \code{\link[BASiCS]{BASiCS_Summary-class}}.
}
\description{
For each of the BASiCS parameters (see Vallejos et al 2015),
\code{Summary} returns the corresponding postior medians and limits of the high posterior
density interval (probabilty equal to \code{prob}. )
}
\examples{
# See documentation of function BASiCS_MCMC
}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}
\references{
Vallejos, Marioni and Richardson (2015). Bayesian Analysis of Single-Cell Sequencing data.
}

