% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASiCS_D_TestDE.R
\name{BASiCS-defunct}
\alias{BASiCS-defunct}
\alias{BASiCS_D_TestDE}
\title{Defunct functions in package \sQuote{BASiCS}}
\usage{
BASiCS_D_TestDE(...)
}
\arguments{
\item{...}{Unused.}
}
\description{
These functions are defunct and no longer available.
}
\details{
These functions are defunct and no longer available. 

  \itemize{
  \item{BASiCS_D_TestDE: \code{\link{BASiCS_TestDE}}}
  }
}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}
}
