% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\docType{methods}
\name{BASiCS_Chain-methods}
\alias{BASiCS_Chain-methods}
\alias{updateObject,BASiCS_Chain-method}
\alias{BASiCS_Chain-methods}
\alias{show,BASiCS_Chain-method}
\title{'updateObject' method for BASiCS_Chain objects}
\usage{
\S4method{updateObject}{BASiCS_Chain}(object)

\S4method{show}{BASiCS_Chain}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{BASiCS_Chain}} object.}

\item{object}{A \code{\linkS4class{BASiCS_Chain}} object.}
}
\value{
Returns an updated \code{\linkS4class{BASiCS_Chain}} object that 
contains all model parameters in a single slot object (list).

Prints a summary of the properties of \code{object}.
}
\description{
'updateObject' method for \code{\linkS4class{BASiCS_Chain}} objects.

'show' method for \code{\linkS4class{BASiCS_Chain}} objects.
}
\examples{

# Not run
# New_Chain <- updateObject(Old_Chain)


Data <- makeExampleBASiCS_Data()
Chain <- BASiCS_MCMC(Data, N = 50, Thin = 2, Burn = 2)

}
\references{
Vallejos, Marioni and Richardson (2015). PLoS Computational Biology.

Vallejos, Marioni and Richardson (2015). PLoS Computational Biology.
}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}

Catalina A. Vallejos \email{cnvallej@uc.cl}
}
