% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\docType{methods}
\name{BASiCS_Chain-methods}
\alias{BASiCS_Chain-methods}
\alias{show,BASiCS_Chain-method}
\title{'show' method for BASiCS_Chain objects}
\usage{
\S4method{show}{BASiCS_Chain}(object)
}
\arguments{
\item{object}{A \code{BASiCS_Chain} object.}
}
\value{
Prints a summary of the properties of \code{object}.
}
\description{
'show' method for \code{\link[BASiCS]{BASiCS_Chain-class}} objects.
}
\examples{

Data = makeExampleBASiCS_Data()
MCMC_Output <- BASiCS_MCMC(Data, N = 50, Thin = 2, Burn = 2)

}
\references{
Vallejos, Marioni and Richardson (2015). Bayesian Analysis of Single-Cell Sequencing data.
}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}
}
