% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{BASiCS_Chain}
\alias{BASiCS_Chain}
\alias{BASiCS_Chain-class}
\title{The BASiCS_Chain class}
\description{
Container of an MCMC sample of the BASiCS' model parameters 
(see Vallejos et al, 2015) as generated by the 
function \code{\link[BASiCS]{BASiCS_MCMC}}.
}
\section{Slots}{

\describe{
\item{\code{parameters}}{List of matrices containing MCMC chains for each model parameter. 
\describe{
\item{mu}{MCMC chain for gene-specific mean expression parameters 
\eqn{\mu_i}, biological genes only 
(matrix with \code{q.bio} columns, all elements must be positive numbers)}
\item{delta}{MCMC chain for gene-specific biological over-dispersion 
parameters \eqn{\delta_i}, biological genes only 
(matrix with \code{q.bio} columns, all elements must be positive numbers)}
\item{phi}{MCMC chain for cell-specific mRNA content normalisation parameters 
\eqn{\phi_j} (matrix with \code{n} columns, all elements must be positive 
numbers and the sum of its elements must be equal to \code{n})}
\item{s}{MCMC chain for cell-specific technical normalisation parameters 
\eqn{s_j} (matrix with \code{n} columns, 
all elements must be positive numbers)}
\item{nu}{MCMC chain for cell-specific random effects \eqn{\nu_j}
(matrix with \code{n} columns, all elements must be positive numbers)}
\item{theta}{MCMC chain for technical over-dispersion parameter(s) 
\eqn{\theta} (matrix, all elements must be positive, 
each colum represents 1 batch)}
}}
}}

\examples{

# A BASiCS_Chain object created by the BASiCS_MCMC function.
Data <- makeExampleBASiCS_Data()
Chain <- BASiCS_MCMC(Data, N = 100, Thin = 2, Burn = 2)

}
\references{
Vallejos, Marioni and Richardson (2015). PLoS Computational Biology.
}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}
}
