% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Methods.R
\docType{methods}
\name{BASiCS_DV_Chain-methods}
\alias{BASiCS_DV_Chain-methods}
\alias{show,BASiCS_DV_Chain-method}
\title{'show' method for BASiCS_DV_Chain objects}
\usage{
\S4method{show}{BASiCS_DV_Chain}(object)
}
\arguments{
\item{object}{A \code{BASiCS_DV_Chain} object.}
}
\value{
Prints a summary of the properties of \code{object}.
}
\description{
'show' method for \code{\link[BASiCSDV]{BASiCS_DV_Chain-class}} objects.
}
\examples{
Data = makeExampleBASiCS_DV_Data()
#MCMC_Output <- BASiCS_DV_MCMC(Data, N = 50, Thin = 2, Burn = 2)
}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}

