% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsPart2.R
\name{BASiCS_D_TestDE}
\alias{BASiCS_D_TestDE}
\title{Detection of genes with changes in expression}
\usage{
BASiCS_D_TestDE(Data, object, GeneNames, EpsilonM = 0.1, EpsilonD = 0.1,
  EviThresholdM = NULL, EviThresholdD = NULL,
  OrderVariable = "ProbDiffExp", GroupLabelRef = "Ref",
  GroupLabelTest = "Test", Plot = FALSE, OffSet = FALSE, EFDR_M = 0.05,
  EFDR_D = 0.05, GenesSelect = NULL, ...)
}
\arguments{
\item{Data}{an object of class \code{\link[BASiCS]{BASiCS_D_Data-class}}}

\item{object}{an object of class \code{\link[BASiCS]{BASiCS_D_Chain-class}}}

\item{GeneNames}{Vector containing gene names to be used in results table (argument to be removed as 'GeneNames' will be an slot of `BASiCS_D_Data` object)}

\item{EpsilonM}{Minimum fold change tolerance threshold for detecting changes in overall expression (must be a positive real number)}

\item{EpsilonD}{Minimum fold change tolerance threshold for detecting changes in cell-to-cell biological over dispersion (must be a positive real number)}

\item{EviThresholdM}{Optional parameter. Evidence threshold for detecting changes in overall expression (must be a positive value, between 0 and 1)}

\item{EviThresholdD}{Optional parameter. Evidence threshold for detecting changes in cell-to-cell biological over dispersion (must be a positive value, between 0 and 1)}

\item{OrderVariable}{Ordering variable for output. Must take values in \code{c("GeneIndex", "GeneNames", "ProbDiffExp", "ProbDiffOverDisp")}.}

\item{GroupLabelRef}{Label assigned to reference group. Default: \code{GroupLabelRef = "Ref"}}

\item{GroupLabelTest}{Label assigned to reference group. Default: \code{GroupLabelRef = "Test"}}

\item{Plot}{If \code{Plot = T}, error rates control rates and volcano plots are generated.}

\item{OffSet}{Optional argument to remove a fix offset effect (if not previously removed from the MCMC chains). This argument will be removed shorly, once offset removal is built as an internal step.}

\item{EFDR_M}{Target for expected false discovery rate related to the comparison of means (default = 0.05)}

\item{EFDR_D}{Target for expected false discovery rate related to the comparison of dispersions (default = 0.05)}

\item{GenesSelect}{Optional argument to provide a user-defined list of genes to be considered for the comparison (default = NULL). When used, this argument must be a vector of 'TRUE' (include gene) / 'FALSE' (exclude gene) indicator, with the same length as the number of intrinsic genes and following the same order as how genes are displayed in the table of counts.  This argument is necessary in order to have a meaningful EFDR calibration when the user decides to exclude some genes from the comparison.}

\item{...}{Graphical parameters (see \code{\link[graphics]{par}}).}
}
\value{
\code{BASiCS_D_TestDE} returns a list of 4 elements:
\describe{
\item{\code{Table}}{A \code{\link[base]{data.frame}} containing the results of the differential expression test}
   \describe{
   \item{\code{Mu}}{Vector of length \code{q.bio}. For each biological gene, posterior median of gene-specific expression levels \eqn{\mu[i]}}
   \item{\code{Delta}}{Vector of length \code{q.bio}. For each biological gene, posterior median of gene-specific biological cell-to-cell heterogeneity hyper-parameter \eqn{\delta[i]}}
   \item{\code{Sigma}}{Vector of length \code{q.bio}. For each biological gene, proportion of the total variability that is due to a cell-to-cell biological heterogeneity component. }
   \item{\code{Prob}}{Vector of length \code{q.bio}. For each biological gene, probability of being highly variable according to the given thresholds.}
   \item{\code{HVG}}{Vector of length \code{q.bio}. For each biological gene, indicator of being detected as highly variable according to the given thresholds. }
   }
\item{\code{EviThreshold}}{Evidence thresholds.}
\item{\code{EFDR}}{Expected false discovery rate for the given thresholds.}
\item{\code{EFNR}}{Expected false negative rate for the given thresholds.}
}
}
\description{
Function to assess changes in expression (mean and over-dispersion)
}
\details{
See vignette
}
\examples{

# See
help(BASiCS_D_MCMC)

}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}
\seealso{
\code{\link[BASiCS]{BASiCS_D_Chain-class}}
}

