% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Methods.R
\docType{methods}
\name{BASiCS_Data-methods}
\alias{BASiCS_Data-methods}
\alias{counts}
\alias{counts,BASiCS_Data-method}
\alias{displaySpikeInput}
\alias{displaySpikeInput,BASiCS_Data-method}
\alias{displayTechIndicator}
\alias{displayTechIndicator,BASiCS_Data-method}
\alias{show,BASiCS_Data-method}
\title{S4 methods for BASiCS_Data objects}
\usage{
\S4method{show}{BASiCS_Data}(object)

\S4method{counts}{BASiCS_Data}(object, type = "all")

\S4method{displaySpikeInput}{BASiCS_Data}(object)

\S4method{displayTechIndicator}{BASiCS_Data}(object)
}
\arguments{
\item{object}{A \code{BASiCS_Data} object.}

\item{type}{Only required for \code{counts} method. A string indicating which genes are returned.}
}
\value{
\describe{
\item{\code{show}}{Prints a summary of the properties of \code{object}.}

\item{\code{counts}}{\describe{
 \item{\code{if(type = "all")}}{Returns the \code{Counts} slot of \code{object}.}
 \item{\code{if(type = "biological")}}{Returns the \code{Counts} slot of \code{object}, biological genes only.}
 \item{\code{if(type = "technical")}}{Returns the \code{Counts} slot of \code{object}, tehcnical genes only.}}}

\item{\code{displayTechIndicator}}{Returns \code{Tech} slot of \code{object}.}

\item{\code{displaySpikeInput}}{Returns \code{SpikeInput} slot of \code{object}.}
}
}
\description{
S4 methods for \code{\link[BASiCS]{BASiCS_Data-class}} objects.
}
\examples{
Data = makeExampleBASiCS_Data()
show(Data)
head(counts(Data))
dim(counts(Data, type="biological"))
dim(counts(Data, type="technical"))
displayTechIndicator(Data)
displaySpikeInput(Data)
}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}
\references{
Vallejos, Marioni and Richardson (2015). Bayesian Analysis of Single-Cell Sequencing data.
}
\seealso{
\code{\link[BASiCS]{BASiCS_Data-class}}
}

