% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASiCS_DenoisedCounts.R
\name{BASiCS_DenoisedCounts}
\alias{BASiCS_DenoisedCounts}
\title{Calculates denoised expression expression counts}
\usage{
BASiCS_DenoisedCounts(Data, Chain)
}
\arguments{
\item{Data}{an object of class \code{\link[SingleCellExperiment]{SingleCellExperiment}}}

\item{Chain}{an object of class \code{\link[BASiCS]{BASiCS_Chain-class}}}
}
\value{
A matrix of denoised expression counts.
}
\description{
Calculates denoised expression counts by removing cell-specific technical variation.
}
\details{
See vignette
}
\examples{

# See
help(BASiCS_MCMC)

}
\references{
Vallejos, Marioni and Richardson (2015). Bayesian Analysis of Single-Cell Sequencing data. PLoS Computational Biology.
}
\seealso{
\code{\link[BASiCS]{BASiCS_Chain-class}}
}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}
}
