% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASiCS_DenoisedCounts.R
\name{BASiCS_DenoisedCounts}
\alias{BASiCS_DenoisedCounts}
\title{Calculates denoised expression expression counts}
\usage{
BASiCS_DenoisedCounts(Data, Chain)
}
\arguments{
\item{Data}{an object of class \code{\linkS4class{SingleCellExperiment}}}

\item{Chain}{an object of class \code{\linkS4class{BASiCS_Chain}}}
}
\value{
A matrix of denoised expression counts. In line with global scaling
normalisation strategies, these are defined as \eqn{X_{ij}/(\phi_j \nu_j)} 
for biological genes and \eqn{X_{ij}/(\nu_j)} for spike-in genes. For this
calculation \eqn{\phi_j} \eqn{\nu_j} are estimated by their corresponding
posterior medians.
}
\description{
Calculates denoised expression counts by removing 
cell-specific technical variation.
}
\details{
See vignette
}
\examples{

# See
help(BASiCS_MCMC)

}
\references{
Vallejos, Marioni and Richardson (2015). PLoS Computational Biology.
}
\seealso{
\code{\linkS4class{BASiCS_Chain}}
}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}
}
