% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASiCS_DenoisedRates.R
\name{BASiCS_DenoisedRates}
\alias{BASiCS_DenoisedRates}
\title{Calculates denoised expression rates}
\usage{
BASiCS_DenoisedRates(Data, Chain, PrintProgress = FALSE,
  Propensities = FALSE)
}
\arguments{
\item{Data}{an object of class \code{\link[SingleCellExperiment]{SingleCellExperiment}}}

\item{Chain}{an object of class \code{\link[BASiCS]{BASiCS_Chain}}}

\item{PrintProgress}{If \code{TRUE}, partial progress information is printed in the console.}

\item{Propensities}{If \code{TRUE}, returns underlying expression propensitites \eqn{\rho_{ij}}. 
Otherwise, denoised rates \eqn{\mu_i \rho_{ij}} are returned.}
}
\value{
A matrix of denoised expression rates.
}
\description{
Calculates normalised and denoised expression rates, by removing the effect of technical variation.
}
\details{
See vignette
}
\examples{

# See
help(BASiCS_MCMC)

}
\references{
Vallejos, Marioni and Richardson (2015). Bayesian Analysis of Single-Cell Sequencing data. PLoS Computational Biology.
}
\seealso{
\code{\link[BASiCS]{BASiCS_Chain}}
}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}
}
