% Generated by roxygen2 (4.0.2): do not edit by hand
\name{BASiCS_DetectHVG}
\alias{BASiCS_DetectHVG}
\alias{BASiCS_DetectHVG_LVG}
\alias{BASiCS_DetectLVG}
\title{Detection method for highly and lowly variable genes}
\usage{
BASiCS_DetectHVG(object, VarThreshold, EviThreshold = NULL,
  OrderVariable = "Prob", Plot = FALSE, ...)

BASiCS_DetectLVG(object, VarThreshold, EviThreshold = NULL,
  OrderVariable = "Prob", Plot = FALSE, ...)
}
\arguments{
\item{object}{an object of class \code{\link[BASiCS]{BASiCS_Chain-class}}}

\item{VarThreshold}{Variance contribution threshold (must be a positive value, between 0 and 1)}

\item{EviThreshold}{Optional parameter. Evidence threshold (must be a positive value, between 0 and 1)}

\item{OrderVariable}{Ordering variable for output. Must take values in \code{c("GeneIndex", "Mu", "Delta", "Sigma", "Prob")}.}

\item{Plot}{If \code{Plot = T} a plot of the gene specific expression level against HVG or LVG is generated.}

\item{...}{Graphical parameters (see \code{\link[graphics]{par}}).}
}
\value{
\code{BASiCS_DetectHVG} returns a list of 4 elements:
\describe{
\item{\code{Table}}{Matrix whose columns contain}
   \describe{
   \item{\code{GeneIndex}}{Gene index (as in input dataset)}
   \item{\code{Mu}}{Vector of length \code{q.bio}. For each biological gene, posterior median of gene-specific expression levels \eqn{\mu[i]}}
   \item{\code{Delta}}{Vector of length \code{q.bio}. For each biological gene, posterior median of gene-specific biological cell-to-cell heterogeneity hyper-parameter \eqn{\delta[i]}}
   \item{\code{Sigma}}{Vector of length \code{q.bio}. For each biological gene, proportion of the total variability that is due to a cell-to-cell biological heterogeneity component. }
   \item{\code{Prob}}{Vector of length \code{q.bio}. For each biological gene, probability of being highly variable according to the given thresholds.}
   \item{\code{HVG}}{Vector of length \code{q.bio}. For each biological gene, indicator of being detected as highly variable according to the given thresholds. }
   }
\item{\code{EviThreshold}}{Evidence threshold.}
\item{\code{EFDR}}{Expected false discovery rate for the given thresholds.}
\item{\code{EFNR}}{Expected false negative rate for the given thresholds.}
}
\code{BASiCS_DetectLVG} produces a similar output, replacing the element \code{HVG} by \code{LVG}, an indicator of a gene being detected as lowly variable according to the given thresholds.
}
\description{
Detection method for highly and lowly variable genes
}
\details{
See vignette
}
\examples{
# See
help(BASiCS_MCMC)
}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}
\references{
Vallejos, Marioni and Richardson (2015). Bayesian Analysis of Single-Cell Sequencing data.
}
\seealso{
\code{\link[BASiCS]{BASiCS_Chain-class}}
}

