% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASiCS_DiagHist.R
\name{BASiCS_DiagHist}
\alias{BASiCS_DiagHist}
\alias{BASiCS_diagHist}
\title{Create diagnostic plots of MCMC parameters}
\usage{
BASiCS_DiagHist(object, Param = NULL, na.rm = TRUE)

BASiCS_diagHist(object, Param = NULL, na.rm = TRUE)
}
\arguments{
\item{object}{an object of class \code{\linkS4class{BASiCS_Summary}}}

\item{Param}{Optional name of a chain parameter to restrict the histogram;
if not supplied, all parameters will be assessed.
Possible values: \code{'mu'}, \code{'delta'}, \code{'phi'},
\code{'s'}, \code{'nu'}, \code{'theta'}, \code{'beta'},
\code{'sigma2'} and \code{'epsilon'}. Default \code{Param = NULL}.}

\item{na.rm}{Logical value indicating whether NA values should be removed
before calculating effective sample size.}
}
\value{
A ggplot object.
}
\description{
Plot a histogram of effective sample size or Geweke's diagnostic
z-statistic. See \link[coda]{effectiveSize} and \link[coda]{geweke.diag} for
more details.
}
\examples{

# Built-in example chain
data(ChainSC)

# See effective sample size distribution across all parameters
BASiCS_DiagHist(ChainSC)
# For mu only
BASiCS_DiagHist(ChainSC, Param = "mu")

}
\seealso{
\code{\linkS4class{BASiCS_Chain}}
}
\author{
Alan O'Callaghan \email{a.b.ocallaghan@sms.ed.ac.uk}
}
