% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASiCS_diagHist.R
\name{BASiCS_diagHist}
\alias{BASiCS_diagHist}
\title{Create diagnostic plots of MCMC parameters}
\usage{
BASiCS_diagHist(object, Param = NULL, na.rm = FALSE)
}
\arguments{
\item{object}{an object of class \code{\linkS4class{BASiCS_Summary}}}

\item{Param}{Optional name of a chain parameter to restrict the histogram;
if not supplied, all parameters will be assessed.
Possible values: \code{'mu'}, \code{'delta'}, \code{'phi'},
\code{'s'}, \code{'nu'}, \code{'theta'}, \code{'beta'},
\code{'sigma2'} and \code{'epsilon'}.}

\item{na.rm}{Logical value indicating whether NA values should be removed
before calculating effective sample size.}
}
\value{
A ggplot object.
}
\description{
Plot a histogram of effective sample size or Geweke's diagnostic
z-statistic. See \link[coda]{effectiveSize} and \link[coda]{geweke.diag} for
more details.
}
\examples{

# See help(BASiCS_MCMC)

# Built-in simulated dataset
Data <- makeExampleBASiCS_Data()
# To analyse real data, please refer to the instructions in:
# https://github.com/catavallejos/BASiCS/wiki/2.-Input-preparation

# Only a short run of the MCMC algorithm for illustration purposes
# Longer runs migth be required to reach convergence
Chain <- BASiCS_MCMC(Data, N = 10, Thin = 2, Burn = 4, Regression = FALSE,
                     PrintProgress = FALSE, WithSpikes = TRUE)

}
\seealso{
\code{\linkS4class{BASiCS_Chain}}
}
\author{
Alan O'Callaghan \email{a.b.ocallaghan@sms.ed.ac.uk}
}
