% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_DivideAndConquer.R
\name{BASiCS_DivideAndConquer}
\alias{BASiCS_DivideAndConquer}
\title{Run divide and conquer MCMC with BASiCS}
\usage{
BASiCS_DivideAndConquer(
  Data,
  NSubsets = 5,
  SubsetBy = c("cell", "gene"),
  Alpha = 0.05,
  WithSpikes = FALSE,
  Regression = TRUE,
  mc.cores = min(detectCores(all.tests = FALSE, logical = TRUE), NSubsets),
  ...
)
}
\arguments{
\item{Data}{SingleCellExperiemnt object}

\item{NSubsets}{Number of partitions to make}

\item{SubsetBy}{Passed to generateSubsets}

\item{Alpha}{Passed to generateSubsets}

\item{WithSpikes}{Similar to the argument in BASiCS_MCMC}

\item{mc.cores}{Number of cores for parallel processing}

\item{...}{Passed to BASiCS_MCMC}
}
\description{
Run divide and conquer MCMC with BASiCS
}
