% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Methods.R
\docType{methods}
\name{BASiCS_DV_Summary-methods}
\alias{BASiCS_DV_Summary-methods}
\alias{show,BASiCS_DV_Summary-method}
\title{'show' method for BASiCS_DV_Summary objects}
\usage{
\S4method{show}{BASiCS_DV_Summary}(object)
}
\arguments{
\item{object}{A \code{BASiCS_DV_Summary} object.}
}
\value{
Prints a summary of the properties of \code{object}.
}
\description{
'show' method for \code{\link[BASiCSDV]{BASiCS_DV_Summary-class}} objects.
}
\examples{
# see help(BASiCS_DV_MCMC)
}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}

