% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASiCS_effectiveSize.R
\name{BASiCS_effectiveSize}
\alias{BASiCS_effectiveSize}
\title{Calculate effective sample size for BASiCS_Chain parameters}
\usage{
BASiCS_effectiveSize(object, Param, na.rm = TRUE)
}
\arguments{
\item{object}{A BASiCS_Chain object.}

\item{Param}{The parameter to use to calculate effectiveSize.}

\item{na.rm}{Remove NA values before calculating effectiveSize. Default is TRUE.}
}
\description{
Calculate effective sample size for BASiCS_Chain parameters
}
\examples{
data(ChainSCReg)
# Will often fail for real chains, when epsilon is NA
effectiveSize(ChainSCReg@parameters[["epsilon"]])
# This has an na.rm argument which removes these NA values before
# calculating effectiveSize
BASiCS_effectiveSize(ChainSC, Param = "mu")

}
