% Generated by roxygen2 (4.0.2): do not edit by hand
\name{BASiCS_MCMC}
\alias{BASiCS_MCMC}
\title{BASiCS MCMC sampler}
\usage{
BASiCS_MCMC(Data, N, Thin, Burn, PriorParam = list(a.delta = 1, b.delta = 1,
  s2.kappa = 1, a.s = 1, b.s = 1, a.theta = 1, b.theta = 1), AR = 0.44,
  StopAdapt = Burn, StoreChains = F, StoreAdapt = F, StoreDir = getwd(),
  RunName = "")
}
\arguments{
\item{Data}{A \code{\link[BASiCS]{BASiCS_Data-class}} object.}

\item{N}{Total number of iterations for the MCMC sampler. Use \code{N>=max(4,Thin)}, \code{N} being a multiple of \code{Thin}.}

\item{Thin}{Thining period for the MCMC sampler. Use \code{Thin>=2}.}

\item{Burn}{Burn-in period for the MCMC sampler. Use \code{Burn>=1}, \code{Burn<N}, \code{Burn} being a multiple of \code{Thin}.}

\item{PriorParam}{List of 7 elements, containing the hyper-parameter values required for the adopted prior (see Vallejos et al, 2015). All elements must be positive real numbers.
\describe{
  \item{\code{a.delta}}{Shape hyper-parameter for the Gamma(\code{a.delta},\code{b.delta}) prior that is shared by all gene-specific biological cell-to-cell heterogeneity hyper-parameters \eqn{\delta[i]}.
  Default: \code{a.delta = 1}.}
  \item{\code{b.delta}}{Rate hyper-parameter for the Gamma(\code{a.delta},\code{b.delta}) prior that is shared by all gene-specific biological cell-to-cell heterogeneity hyper-parameters \eqn{\delta[i]}.
  Default: \code{b.delta = 1}.}
  \item{\code{s2.kappa}}{Variance hyper-parameter for the Normal(0,\code{s2.kappa}) prior that is shared by all cell-specific mRNA content normalising constants \eqn{\kappa[j]} (logit scale).
  Default: \code{s2.kappa = 1}.}
  \item{\code{a.s}}{Shape hyper-parameter for the Gamma(\code{a.s},\code{b.s}) prior that is shared by all cell-specific capture efficiency normalising constants \eqn{\s[j]}.
  Default: \code{a.s = 1}.}
  \item{\code{b.s}}{Rate hyper-parameter for the Gamma(\code{a.s},\code{b.s}) prior that is shared by all cell-specific capture efficiency normalising constants \eqn{\s[j]}.
  Default: \code{b.s = 1}.}
  \item{\code{a.theta}}{Shape hyper-parameter for the Gamma(\code{a.theta},\code{b.theta}) prior for technical noise hyper-parameter \eqn{\theta}.
  Default: \code{a.theta = 1}.}
  \item{\code{b.theta}}{Rate hyper-parameter for the Gamma(\code{a.theta},\code{b.theta}) prior for technical noise hyper-parameter \eqn{\theta}.
  Default: \code{b.theta = 1}.}
}}

\item{AR}{Optimal acceptance rate for adaptive Metropolis Hastings updates. It must be a positive number between 0 and 1. Default (and recommended): \code{ar = 0.44}.}

\item{StopAdapt}{Iteration at which adaptive proposals are not longer adapted. Use \code{stopAdapt>=1}. Default: \code{stopAdapt = burn}.}

\item{StoreChains}{If \code{StoreChains = T}, MCMC chains of each parameter are stored in separate .txt files. (\code{RunName} argument used for file names). Default: \code{StoreChains = F}.}

\item{StoreAdapt}{If \code{StoreAdapt = T}, trajectory of adaptive proposal variances (log scale) for each parameter
are stored in separate .txt files. (\code{RunName} argument used for file names). Default: \code{StoreAdapt = F}.}

\item{StoreDir}{Directory where MCMC chain will be stored (only required if \code{Store = T}). Default: \code{StoreDir = getwd()}.}

\item{RunName}{Run-name to be used when storing chains and/or adaptive proposal variances in .txt files.}
}
\value{
An object of class \code{\link[BASiCS]{BASiCS_Chain-class}}.
}
\description{
MCMC sampler to perform Bayesian inference for single-cell mRNA sequencing datasets using the model described in Vallejos et al (2015).
}
\examples{
Data = makeExampleBASiCS_Data()

# Only a short run of the MCMC algorithm for illustration purposes
# Longer runs required to reach convergence
Result <- BASiCS_MCMC(Data, N = 100, Thin = 2, Burn = 2)

Summary(Result)

plot(Result, Param = "mu", Column = 1)
plot(Result, Param = "delta", Column = 1)
plot(Result, Param = "phi", Column = 1)
plot(Result, Param = "s", Column = 1)
plot(Result, Param = "nu", Column = 1)
plot(Result, Param = "theta")
}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}
\references{
Vallejos, Marioni and Richardson (2015). Bayesian Analysis of Single-Cell Sequencing data.
}

