% Generated by roxygen2 (4.0.2): do not edit by hand
\name{BASiCS_MCMC_Start}
\alias{BASiCS_MCMC_Start}
\title{Starting values for BASiCS_MCMC}
\usage{
BASiCS_MCMC_Start(Data, ...)
}
\arguments{
\item{Data}{A \code{\link[BASiCS]{BASiCS_Data-class}} object.}

\item{...}{Optional arguments (lower bounds for starting values of adaptive proposal variances, log-scale):
\describe{
 \item{\code{ls.mu0}}{Related to gene-specific expression levels \eqn{\mu[i]}.}
 \item{\code{ls.delta0}}{Related to gene-specific biological cell-to-cell heterogeneity hyper-parameters \eqn{\delta[i]}.}
 \item{\code{ls.kappa0}}{Related to cell-specific mRNA content normalising constants (logit-scale) \eqn{\kappa[j]}}
 \item{\code{ls.nu0}}{Related to cell-specific random effects (technical noise) \eqn{\nu[j]}.}
 \item{\code{ls.theta0}}{Related to technical variability hyper-parameter \eqn{\theta}.}
}}
}
\value{
A list containing starting values for all model parameters and the corresponding variances of the adaptive proposals (log-scale).
}
\description{
Auxiliary function that creates starting values for \code{\link[BASiCS]{BASiCS_MCMC}}.
}
\details{
Typically, this function will not be directly called. It is internally used by \code{\link[BASiCS]{BASiCS_MCMC}}.
}
\examples{
Data = makeExampleBASiCS_Data()

# Creating starting values for BASiCS_MCMC
MCMC_Start <- BASiCS_MCMC_Start(Data)

# Creating starting values for BASiCS_MCMC
# With user-defined lower bounds for adaptive proposal variances of gene-specific expression rates
MCMC_Start2 <- BASiCS_MCMC_Start(Data, ls.mu0 = -6)
}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}
\references{
Vallejos, Marioni and Richardson (2015). Bayesian Analysis of Single-Cell Sequencing data.
}

