% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BASiCS_OffsetCorrected}
\alias{BASiCS_OffsetCorrected}
\alias{BASiCS_OffsetCorrected-class}
\title{The BASiCS_OffsetCorrected class}
\description{
Container of two chains which have been offset to have matching
mean mu values.
}
\section{Slots}{

\describe{
\item{\code{Chain1,Chain2}}{Objects of class \code{\linkS4class{BASiCS_Chain}}.}

\item{\code{GroupLabel1,GroupLabel2}}{Labels for Chain1 and Chain2, respectively.}

\item{\code{OffsetChain}}{Ratio between sum of mus}

\item{\code{OffsetEst}}{Median offset estimate}

\item{\code{ChainTau}}{log2 fold-change between MCMC draws for mu}

\item{\code{Mu1}}{Median mu values for Chain1}

\item{\code{Mu1_old}}{Median mu values for Chain1 before scaling}

\item{\code{Mu2}}{Median mu values for Chain2}

\item{\code{MuBase}}{(Mu1 * n1 + Mu2 * n2) / n}

\item{\code{MuBaseOld}}{(Mu1 * n1 + Mu2 * n2) / n before offset correction}

\item{\code{MedianTau}}{Median fold change}

\item{\code{MedianTau_old}}{Median fold change before offset correction}

\item{\code{Delta1,Delta2}}{Median delta for chain 1 and chain 2}
}}

