% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BASiCS_OffsetCorrected}
\alias{BASiCS_OffsetCorrected}
\alias{BASiCS_OffsetCorrected-class}
\title{The BASiCS_OffsetCorrected class}
\description{
Container of two chains which have been offset to have matching
mean mu values.
}
\section{Slots}{

\describe{
\item{\code{Chain1,Chain2}}{Objects of class \code{\linkS4class{BASiCS_Chain}}.}

\item{\code{GroupLabel1,GroupLabel2}}{Labels for Chain1 and Chain2, respectively.}

\item{\code{OffsetChain}}{}
}}

