% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASiCS_PlotDE.R
\name{BASiCS_PlotDE,BASiCS_ResultsDE-method}
\alias{BASiCS_PlotDE,BASiCS_ResultsDE-method}
\title{Produce plots assessing differential expression results}
\usage{
\S4method{BASiCS_PlotDE}{BASiCS_ResultsDE}(
  object,
  Plots = c("MAPlot", "VolcanoPlot", "GridPlot"),
  Parameters = intersect(c("Mean", "Disp", "ResDisp"), names(object@Results)),
  MuX = TRUE,
  ...
)
}
\arguments{
\item{object}{A BASiCS_ResultsDE or BASiCS_ResultDE object.}

\item{Plots}{Plots plot to produce? Options: "MAPlot", "VolcanoPlot", 
"GridPlot"}

\item{Parameters}{Which parameter(s) to produce plots for?
Available options are "Mean", (mu mean expression),
"Disp" (delta overdispersion) and "ResDisp" 
(epsilon residual overdispersion).}

\item{MuX}{Use Mu (mean expression across both chains) as the X-axis for all
MA plots? Default: TRUE.}

\item{...}{Passed to methods}
}
\value{
Plot objects.
}
\description{
Produce plots assessing differential expression results
}
\examples{
data(ChainSC)
data(ChainRNA)

Test <- BASiCS_TestDE(Chain1 = ChainSC, Chain2 = ChainRNA,
                      GroupLabel1 = 'SC', GroupLabel2 = 'P&S',
                      EpsilonM = log2(1.5), EpsilonD = log2(1.5),
                      OffSet = TRUE)
BASiCS_PlotDE(Test)
}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}

Nils Eling \email{eling@ebi.ac.uk}

Alan O'Callaghan \email{a.b.o'callaghan@sms.ed.ac.uk}
}
