% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASiCS_PlotDE.R
\docType{methods}
\name{BASiCS_PlotDE,BASiCS_ResultsDE-method}
\alias{BASiCS_PlotDE,BASiCS_ResultsDE-method}
\title{Produce plots assessing differential expression results}
\usage{
\S4method{BASiCS_PlotDE}{BASiCS_ResultsDE}(object, Which = c("MAPlot",
  "VolcanoPlot", "GridPlot"), ...)
}
\arguments{
\item{object}{A BASiCS_ResultsDE or BASiCS_ResultDE object.}

\item{Which}{Which plot to produce? Options: "MAPlot", "VolcanoPlot", 
"GridPlot"}

\item{...}{Passed to methods}
}
\description{
Produce plots assessing differential expression results
}
\examples{
data(ChainSC)
data(ChainRNA)

Test <- BASiCS_TestDE(Chain1 = ChainSC, Chain2 = ChainRNA,
                      GroupLabel1 = 'SC', GroupLabel2 = 'P&S',
                      EpsilonM = log2(1.5), EpsilonD = log2(1.5),
                      OffSet = TRUE)
BASiCS_PlotDE(Test)
BASiCS_PlotDE(Test@Results[[1]])
}
