% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASiCS_PlotOffSet.R
\name{BASiCS_PlotOffset}
\alias{BASiCS_PlotOffset}
\title{Visualise global offset in mean expression between two chains.}
\usage{
BASiCS_PlotOffset(OffsetCorrected, Type = c("offset estimate",
  "before-after", "MA plot"), Print = FALSE)
}
\arguments{
\item{OffsetCorrected}{Object of class BASiCS_OffsetCorrected}

\item{Type}{The type of plot generated.
\code{"offset estimate"} produces a boxplot of the offset alongside
an estimate of the global offset.
\code{"before-after"} produces MA plots of Mean expression against
log2(fold-change) before and after offset correction.
\code{"MA plot"} produces an MA plot of Mean expression against 
log2(fold-change).}

\item{Print}{Print the plots (TRUE), or simply return them as a list (FALSE)?
Default is \code{FALSE}.}
}
\description{
Visualise global offset in mean expression between two 
\code{BASiCS_Chain} objects.
}
\examples{

# Loading two 'BASiCS_Chain' objects (obtained using 'BASiCS_MCMC')
data(ChainSC)
data(ChainRNA)

Corrected <- BASiCS_CorrectOffset(ChainSC, ChainRNA, "a", "b", Plot = FALSE)
BASiCS_PlotOffset(Corrected)


}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}

Nils Eling \email{eling@ebi.ac.uk}

Alan O'Callaghan \email{a.b.o'callaghan@sms.ed.ac.uk}
}
