% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASiCS_PriorParam.R
\name{BASiCS_PriorParam}
\alias{BASiCS_PriorParam}
\title{Prior parameters for BASiCS_MCMC}
\usage{
BASiCS_PriorParam(
  Data,
  k = 12,
  mu.mu = 0,
  s2.mu = 0.5,
  s2.delta = 0.5,
  a.delta = 1,
  b.delta = 1,
  p.phi = rep(1, times = ncol(Data)),
  a.s = 1,
  b.s = 1,
  a.theta = 1,
  b.theta = 1,
  RBFMinMax = TRUE,
  FixLocations = !is.null(RBFLocations),
  RBFLocations = NULL,
  m = numeric(k),
  V = diag(k),
  a.sigma2 = 2,
  b.sigma2 = 2,
  eta = 5,
  GeneExponent = 1,
  CellExponent = 1
)
}
\arguments{
\item{Data}{\linkS4class{SingleCellExperiment} object (required).}

\item{k}{Number of regression terms, including k - 2 Gaussian 
radial basis functions (GRBFs).}

\item{mu.mu, s2.mu}{Mean and variance parameters for lognormal prior on mu.}

\item{s2.delta}{Variance parameter for lognormal prior on delta when 
\code{PriorDelta="lognormal"}.}

\item{a.delta, b.delta}{Parameters for gamma prior on delta when 
\code{PriorDelta="gamma"}.}

\item{p.phi}{Parameter for dirichlet prior on phi.}

\item{a.s, b.s}{Parameters for gamma prior on s.}

\item{a.theta, b.theta}{Parameters for gamma prior on theta.}

\item{RBFMinMax}{Should GRBFs be placed at the minimum and maximum of 
\code{log(mu)}?}

\item{FixLocations}{Should RBFLocations be fixed throughout MCMC, or adaptive
during burn-in?}

\item{RBFLocations}{RBFLocations of GRBFs in units of \code{log(mu)}}

\item{m, V}{Mean and (co)variance priors for regression coefficients.}

\item{a.sigma2, b.sigma2}{Priors for inverse gamma prior on regression scale.}

\item{eta}{Degrees of freedom for t distribution of regresion errors.}

\item{GeneExponent, CellExponent}{Exponents for gene and cell-specific 
parameters. These should not be outside of divide and conquer MCMC 
applications.}
}
\description{
This is a convenience function to allow partial specification of prior 
parameters, and to ensure default parameters are consistent across usage
within the package.
}
\examples{

BASiCS_PriorParam(makeExampleBASiCS_Data(), 12)

}
