% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BASiCS_ResultVG}
\alias{BASiCS_ResultVG}
\alias{BASiCS_ResultVG-class}
\title{The BASiCS_ResultVG class}
\description{
Container of results for a single HVG/LVG test.
}
\section{Slots}{

\describe{
\item{\code{Method}}{Character value detailing whether the test performed using 
variance decomposition (\code{method="Variance"}) or percentiles of epsilon 
(\code{Method="Percentile"}).}

\item{\code{RowData}}{Optional \linkS4class{DataFrame} containing 
additional information about genes used in the test.}
}}

