% Generated by roxygen2 (4.0.2): do not edit by hand
\name{BASiCS_Sim}
\alias{BASiCS_Sim}
\title{Simulates expression counts according to the model implemented in BASiCS}
\usage{
BASiCS_Sim(mu, delta, phi, s, theta)
}
\arguments{
\item{mu}{Gene-specific expression levels \eqn{\mu[i]}, defined as true input molecules in case of technical genes
(vector of length \code{q}, technical genes located at the end of the vector, all elements must be positive numbers)}

\item{delta}{Gene-specific biological cell-to-cell heterogeneity hyper-parameters \eqn{\delta[i]}, biological genes only
(vector of length \code{q.bio}, all elements must be positive numbers)}

\item{phi}{Cell-specific mRNA content normalising constants \eqn{\phi[j]}
(vector of length \code{n}, all elements must be positive numbers and the sum of its elements must be equal to \code{n})}

\item{s}{Cell-specific capture efficiency (or amplification biases if not using UMI based counts) normalising constants \eqn{s[j]}
(vector of length \code{n}, all elements must be positive numbers)}

\item{theta}{Technical variability hyper-parameter \eqn{\theta} (must be positive)}
}
\value{
\code{BASiCS_Sim} returns an object of class \code{\link[BASiCS]{BASiCS_Data-class}}.
}
\description{
\code{BASiCS_Sim} creates a simulated dataset from the model implemented in BASiCS.
This function is used in order to illustrate the performance of the \code{BASiCS} library.
}
\examples{
data(ParamMouseShort)
attach(ParamMouseShort)
Data = BASiCS_Sim(MuShort, DeltaShort, PhiShort, SShort, ThetaShort)
head(counts(Data))
dim(counts(Data, type="biological"))
dim(counts(Data, type="technical"))
displayTechIndicator(Data)
displaySpikeInput(Data)
detach(ParamMouseShort)
}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}
\references{
Vallejos, Marioni and Richardson (2015). Bayesian Analysis of Single-Cell Sequencing data.
}

