% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASiCS_Sim.R
\name{BASiCS_Sim}
\alias{BASiCS_Sim}
\title{Generates synthetic data according to the model implemented in BASiCS}
\usage{
BASiCS_Sim(Mu, Mu_spikes, Delta, Phi, S, Theta)
}
\arguments{
\item{Mu}{Gene-specific mean expression parameters \eqn{\mu_i} for all
biological genes (vector of length \code{q.bio},
all elements must be positive numbers)}

\item{Mu_spikes}{\eqn{\mu_i} for all technical genes defined as true
input molecules (vector of length \code{q-q.bio},
all elements must be positive numbers)}

\item{Delta}{Gene-specific biological over-dispersion parameters
\eqn{\delta_i}, biological genes only
(vector of length \code{q.bio}, all elements must be positive numbers)}

\item{Phi}{Cell-specific mRNA content normalising parameters \eqn{\phi_j}
(vector of length \code{n}, all elements must be positive numbers and
the sum of its elements must be equal to \code{n})}

\item{S}{Cell-specific technical normalising parameters \eqn{s_j}
(vector of length \code{n}, all elements must be positive numbers)}

\item{Theta}{Technical variability parameter \eqn{\theta} (must be positive)}
}
\value{
An object of class \code{\linkS4class{SingleCellExperiment}},
including synthetic data generated by the model implemented in BASiCS.
}
\description{
\code{BASiCS_Sim} creates a simulated dataset from the model
implemented in BASiCS.
}
\examples{

# Simulated parameter values for 10 genes
# (7 biogical and 3 spike-in) measured in 5 cells
Mu <- c(8.36, 10.65, 4.88, 6.29, 21.72, 12.93, 30.19)
Mu_spike <-  c(1010.72, 7.90, 31.59)
Delta <- c(1.29, 0.88, 1.51, 1.49, 0.54, 0.40, 0.85)
Phi <- c(1.00, 1.06, 1.09, 1.05, 0.80)
S <- c(0.38, 0.40, 0.38, 0.39, 0.34)
Theta <- 0.39

Data <- BASiCS_Sim(Mu, Mu_spike, Delta, Phi, S, Theta)
head(SingleCellExperiment::counts(Data))
dim(SingleCellExperiment::counts(Data))
S4Vectors::metadata(Data)$SpikeInput
SingleCellExperiment::isSpike(Data)

}
\references{
Vallejos, Marioni and Richardson (2015). PLoS Computational Biology.
}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}, Nils Eling
}
