% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{BASiCS_Summary-class}
\alias{BASiCS_Summary-class}
\alias{BASiCS_Summary,BASiCS_Summary-class}
\title{The BASiCS_Summary class}
\description{
Container of a summary of a \code{\link[BASiCS]{BASiCS_Chain-class}} object.  
In each slot, first column contains posterior medians; second and third columns respectively contain the 
lower and upper limits of an high posterior density interval (for a given probability).
}
\section{Slots}{

\describe{
\item{\code{mu}}{Posterior medians (first column), lower (second column) and upper (third column) limits of gene-specific expression levels \eqn{\mu_i}.}

\item{\code{delta}}{Posterior medians (first column), lower (second column) and upper (third column) limits of gene-specific biological cell-to-cell over-dispersion \eqn{\delta_i}, biological genes only}

\item{\code{phi}}{Posterior medians (first column), lower (second column) and upper (third column) limits of cell-specific mRNA content normalising constants \eqn{\phi_j}}

\item{\code{s}}{Posterior medians (first column), lower (second column) and upper (third column) limits of cell-specific capture efficiency (or amplification biases if not using UMI based counts) normalising constants \eqn{s[j]}}

\item{\code{nu}}{Posterior medians (first column), lower (second column) and upper (third column) limits of cell-specific random effects \eqn{\nu_j}}

\item{\code{theta}}{Posterior median (first column), lower (second column) and upper (third column) limits of technical variability hyper-parameter \eqn{\theta} (each row represents one batch)}
}}

\examples{

# A BASiCS_Summary object created by the Summary method.
Data = makeExampleBASiCS_Data()
Chain <- BASiCS_MCMC(Data, N = 100, Thin = 2, Burn = 2)
ChainSummary <- Summary(Chain)

}
