% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\docType{methods}
\name{BASiCS_Summary-methods}
\alias{BASiCS_Summary-methods}
\alias{show,BASiCS_Summary-method}
\title{'show' method for BASiCS_Summary objects}
\usage{
\S4method{show}{BASiCS_Summary}(object)
}
\arguments{
\item{object}{A \code{BASiCS_Summary} object.}
}
\value{
Prints a summary of the properties of \code{object}.
}
\description{
'show' method for \code{\link[BASiCS]{BASiCS_Summary-class}} objects.
}
\examples{

# See
help(BASiCS_MCMC)

}
\references{
Vallejos, Marioni and Richardson (2015). Bayesian Analysis of Single-Cell Sequencing data. PLoS Computational Biology.
}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}
}
