% Generated by roxygen2 (4.0.2): do not edit by hand
\name{BASiCS_VarianceDecomp}
\alias{BASiCS_VarianceDecomp}
\title{Decomposition of gene expression variability according to BASiCS}
\usage{
BASiCS_VarianceDecomp(object, OrderVariable = "BioVar")
}
\arguments{
\item{object}{an object of class \code{\link[BASiCS]{BASiCS_Chain-class}}}

\item{OrderVariable}{Ordering variable for output. Must take values in \code{c("GeneIndex", "BioVar", "TechVar", "ShotNoise")}.}
}
\value{
A matrix with 4 columns
\describe{
\item{\code{GeneIndex}}{Gene index (as in input dataset)}
\item{\code{BioVar}}{Percentage of variance explained by a biological cell-to-cell heterogeneity component}
\item{\code{TechVar}}{Percentage of variance explained by the technical cell-to-cell heterogeneity component}
\item{\code{ShotNoise}}{Percentage of variance explained by the shot noise component (baseline)}
}
}
\description{
Decomposition of gene expression variability according to BASiCS
}
\details{
See vignette
}
\examples{
# See
help(BASiCS_MCMC)
}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}
\references{
Vallejos, Marioni and Richardson (2015). Bayesian Analysis of Single-Cell Sequencing data.
}
\seealso{
\code{\link[BASiCS]{BASiCS_Chain-class}}
}

