% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASiCS_VarianceDecomp.R
\name{BASiCS_VarianceDecomp}
\alias{BASiCS_VarianceDecomp}
\title{Decomposition of gene expression variability according to BASiCS}
\usage{
BASiCS_VarianceDecomp(Chain, OrderVariable = "BioVarGlobal", Plot = TRUE,
  ...)
}
\arguments{
\item{Chain}{an object of class \code{\link[BASiCS]{BASiCS_Chain}}}

\item{OrderVariable}{Ordering variable for output. Possible values: \code{'GeneNames'}, 
\code{'BioVarGlobal'},
 \code{'TechVarGlobal'} and \code{'ShotNoiseGlobal'}.}

\item{Plot}{If \code{TRUE}, a barplot of the variance decomposition 
(global and by batches, if any) is generated}

\item{...}{Other arguments to be passed to \code{\link[graphics]{barplot}}}
}
\value{
A \code{\link[base]{data.frame}} whose first 4 columns correspond to
\describe{
\item{\code{GeneName}}{Gene name (as indicated by user)}
\item{\code{BioVarGlobal}}{Percentage of variance explained by a biological 
cell-to-cell heterogeneity component (overall across all cells)}
\item{\code{TechVarGlobal}}{Percentage of variance explained by the technical 
cell-to-cell heterogeneity component (overall across all cells)}
\item{\code{ShotNoiseGlobal}}{Percentage of variance explained by the shot noise 
component (baseline, overall across all cells)}
}
If more than 1 batch of cells are being analysed, the remaining columns contain 
the corresponding variance decomposition calculated within each batch.
}
\description{
Function to decompose total variability of gene 
expression into biological and technical components.
}
\details{
See vignette
}
\examples{

# See
help(BASiCS_MCMC)

}
\references{
Vallejos, Marioni and Richardson (2015). PLoS Computational Biology.
}
\seealso{
\code{\link[BASiCS]{BASiCS_Chain}}
}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}
}
