% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASiCS_VarThresholdSearchHVG_LVG.R
\name{BASiCS_VarThresholdSearchHVG}
\alias{BASiCS_VarThresholdSearchHVG}
\alias{BASiCS_VarThresholdSearchHVG_LVG}
\alias{BASiCS_VarThresholdSearchLVG}
\alias{BASiCS_VarThresholdSearchHVG_LVG}
\title{Detection method for highly and lowly variable genes using a grid of variance contribution thresholds}
\usage{
BASiCS_VarThresholdSearchHVG(Chain, VarThresholdsGrid, EFDR = 0.1,
  Progress = TRUE)

BASiCS_VarThresholdSearchLVG(Chain, VarThresholdsGrid, EFDR = 0.1,
  Progress = TRUE)
}
\arguments{
\item{Chain}{an object of class \code{\link[BASiCS]{BASiCS_Chain-class}}}

\item{VarThresholdsGrid}{Grid of values for the variance contribution threshold (they must be contained in (0,1))}

\item{EFDR}{Target for expected false discovery rate related to HVG/LVG detection (default = 0.10)}

\item{Progress}{If \code{Progress = TRUE}, partial output is printed in the console.}
}
\value{
\describe{
\item{\code{BASiCS_VarThresholdSearchHVG}}{A table displaying the results of highly variable genes detecting for different variance contribution thresholds.}
\item{\code{BASiCS_VarThresholdSearchLVG}}{A table displaying the results of lowly variable genes detecting for different variance contribution thresholds.oo}
}
}
\description{
Detection method for highly and lowly variable genes using a grid of variance contribution thresholds
}
\details{
See vignette
}
\examples{

# See
help(BASiCS_MCMC)

}
\seealso{
\code{\link[BASiCS]{BASiCS_Chain-class}}
}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}
}
