% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ChainRNA}
\alias{ChainRNA}
\title{Extract from the chain Vallejos et al (2016): pool-and-split samples}
\format{\code{\link[BASiCS]{BASiCS_Chain-class}} containing 10 MCMC iterations}
\usage{
ChainRNA
}
\description{
Small extract (50 MCMC iterations, 1000 randomly selected genes) from the chain obtained in Vallejos et al (2016),
related to pool-and-split samples (this corresponds to the RNA 2i samples in Grun et al, 2014).
}
\references{
Vallejos et al (2016) - Genome Biology
Grun et al (2014) - Nature Methods
}
\keyword{datasets}
