% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ChainRNAReg}
\alias{ChainRNAReg}
\title{Extract from the chain obtained from the regression model: pool-and-split samples}
\format{An object of class \code{\linkS4class{BASiCS_Chain}} 
containing 10 MCMC iterations.}
\usage{
ChainRNAReg
}
\description{
Small extract (10 MCMC iterations, 1000 randomly selected genes) from the 
chain obtained by the regression model, related to pool-and-split samples 
(this corresponds to the RNA 2i samples in Grun et al, 2014).
}
\references{
Grun, Kester and van Oudenaarden (2014). Nature Methods.
}
\keyword{datasets}
