% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ChainSC}
\alias{ChainSC}
\title{Extract from the chain Vallejos et al (2016): single-cell samples}
\format{\code{\link[BASiCS]{BASiCS_Chain-class}} containing 10 MCMC iterations}
\usage{
ChainSC
}
\description{
Small extract (50 MCMC iterations, 1000 randomly selected genes) from the chain obtained in Vallejos et al (2016),
related to single-cell samples (this corresponds to the SC 2i samples in Grun et al, 2014).
}
\references{
Vallejos et al (2016) - Genome Biology
Grun et al (2014) - Nature Methods
}
\keyword{datasets}
