% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ChainSC}
\alias{ChainSC}
\title{Extract from the chain obtained in Vallejos et al (2016): single-cell samples}
\format{An object of class \code{\linkS4class{BASiCS_Chain}} 
containing 10 MCMC iterations.}
\usage{
ChainSC
}
\description{
Small extract (100 MCMC iterations, 500 randomly selected genes) from the 
chain obtained in Vallejos et al (2016), related to single-cell samples 
(this corresponds to the SC 2i samples in Grun et al, 2014).
}
\references{
Vallejos, Richardson and Marioni (2016). Genome Biology.

Grun, Kester and van Oudenaarden (2014). Nature Methods.
}
\keyword{datasets}
