% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ChainSC}
\alias{ChainSC}
\title{Extract from the chain obtained for the Grun et al (2014) data:
single-cell samples}
\format{
An object of class \code{\linkS4class{BASiCS_Chain}} 
containing 75 MCMC iterations.
}
\usage{
ChainSC
}
\description{
Small extract (75 MCMC iterations, 350 randomly selected genes) from the 
chain obtained for the pool-and-split samples
(this corresponds to the SC 2i samples in Grun et al, 2014).
}
\references{
Grun, Kester and van Oudenaarden (2014). Nature Methods.
}
\keyword{datasets}
