% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ChainSCReg}
\alias{ChainSCReg}
\title{Extract from the chain obtained from the regression model: single-cell samples}
\format{An object of class \code{\linkS4class{BASiCS_Chain}} 
containing 10 MCMC iterations.}
\usage{
ChainSCReg
}
\description{
Small extract (10 MCMC iterations, 1000 randomly selected genes) from the 
chain obtained by the regression model, related to single-cell samples 
(this corresponds to the SC 2i samples in Grun et al, 2014).
}
\references{
Grun, Kester and van Oudenaarden (2014). Nature Methods.
}
\keyword{datasets}
