% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DenoisedCounts.R
\name{DenoisedCounts}
\alias{DenoisedCounts}
\alias{BASiCS_DenoisedCounts}
\title{Calculates normalised and denoised expression expression counts}
\usage{
BASiCS_DenoisedCounts(Data, Chain)
}
\arguments{
\item{Data}{an object of class \code{\link[SummarizedExperiment]{SummarizedExperiment}}}

\item{Chain}{an object of class \code{\link[BASiCS]{BASiCS_Chain-class}}}
}
\value{
A matrix of normalised and denoised expression counts.
}
\description{
Calculates normalised and denoised expression counts, by removing the effect of technical variation.
}
\details{
See vignette
}
\examples{

# See
help(BASiCS_MCMC)

}
\seealso{
\code{\link[BASiCS]{BASiCS_Chain-class}}
}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}
}
