% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DenoisedRates.R
\name{DenoisedRates}
\alias{DenoisedRates}
\alias{BASiCS_DenoisedRates}
\title{Calculates normalised and denoised expression rates}
\usage{
BASiCS_DenoisedRates(Data, Chain, PrintProgress = FALSE,
  Propensities = FALSE)
}
\arguments{
\item{Data}{an object of class \code{\link[SummarizedExperiment]{SummarizedExperiment}}}

\item{Chain}{an object of class \code{\link[BASiCS]{BASiCS_Chain-class}}}

\item{PrintProgress}{If \code{TRUE}, partial progress information is printed in the console.}

\item{Propensities}{If \code{TRUE}, returns underlying expression propensitites \eqn{\rho[ij]}. Otherwise, denoised rates \eqn{\mu[i] \rho[ij]} are returned.}
}
\value{
A matrix of normalised and denoised expression counts.
}
\description{
Calculates normalised and denoised expression rates, by removing the effect of technical variation.
}
\details{
See vignette
}
\examples{

# See
help(BASiCS_MCMC)

}
\seealso{
\code{\link[BASiCS]{BASiCS_Chain-class}}
}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}
}
