% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\docType{methods}
\name{displaySummaryBASiCS-BASiCS_Summary-method}
\alias{displaySummaryBASiCS-BASiCS_Summary-method}
\alias{displaySummaryBASiCS}
\alias{displaySummaryBASiCS,BASiCS_Summary-method}
\title{Accessors for the slots of a \code{\linkS4class{BASiCS_Summary}} object}
\usage{
\S4method{displaySummaryBASiCS}{BASiCS_Summary}(object, Param = "mu")
}
\arguments{
\item{object}{an object of class \code{\linkS4class{BASiCS_Summary}}}

\item{Param}{Name of the slot to be used for the accessed. 
Possible values: \code{'mu'}, \code{'delta'}, \code{'phi'}, 
\code{'s'}, \code{'nu'}, \code{'theta'}, \code{'beta'}, 
\code{'sigma2'} and \code{'epsilon'}.}
}
\value{
The requested slot of a \code{\linkS4class{BASiCS_Summary}} object
}
\description{
Accessors for the slots of a \code{\linkS4class{BASiCS_Summary}}
object
}
\examples{

# See
help(BASiCS_MCMC)
  
}
\seealso{
\code{\linkS4class{BASiCS_Summary}}
}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}

Nils Eling \email{eling@ebi.ac.uk}
}
