% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeExampleBASiCS_Data.R
\name{makeExampleBASiCS_Data}
\alias{makeExampleBASiCS_Data}
\title{Create a simple example of a SummarizedExperiment object with random data}
\usage{
makeExampleBASiCS_Data(WithBatch = FALSE, WithSpikes = TRUE, Case1 = TRUE)
}
\arguments{
\item{WithBatch}{If true, 2 batches are generated (each of them containing 10 cells)}

\item{WithSpikes}{If true, the simulated dataset contains 20 spike-in genes}

\item{Case1}{If true, a cell population containing the first 50 biological genes is generates. If false, the second set of 50 biological genes is used to simulated the population.
This is used to generate two populations for differential testing.}
}
\value{
An object of class \code{\link[SummarizedExperiment]{SummarizedExperiment}}, simulated from the model implemented in BASiCS.
It contains 70 genes (50 biological and 20 spike-in) and 20 cells.
}
\description{
A simple \code{\link[SummarizedExperiment]{SummarizedExperiment}} object is generated by simulating a dataset from the
model in BASiCS (Vallejos et al 2015). This is used for the examples in the package and the vignette.
}
\examples{
Data = makeExampleBASiCS_Data()
is(Data, "SummarizedExperiment")

}
\references{
Vallejos, Marioni and Richardson (2015). Bayesian Analysis of Single-Cell Sequencing data.
}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}
}
