% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeExampleBASiCS_Data}
\alias{makeExampleBASiCS_Data}
\title{Create a simple example of a BASiCS_Data object with random data}
\usage{
makeExampleBASiCS_Data()
}
\value{
The simulated dataset contains 100 biological genes and 46 spike-in genes, in 41 cells.
}
\description{
A simple \code{\link[BASiCS]{BASiCS_Data-class}} object is generated by simulating a dataset from the
model in BASiCS (Vallejos et al 2015). This is used for the examples in the paper.
}
\examples{
Data = makeExampleBASiCS_Data()
is(Data, "BASiCS_Data")
}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}
\references{
Vallejos, Marioni and Richardson (2015). Bayesian Analysis of Single-Cell Sequencing data.
}

