% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsPart1.R
\name{makeExampleBASiCS_Data}
\alias{makeExampleBASiCS_Data}
\title{Create a simple example of a BASiCS_Data object with random data}
\usage{
makeExampleBASiCS_Data(WithBatch = FALSE, WithSpikes = TRUE)
}
\arguments{
\item{WithBatch}{If true, 2 batches are generated (each of them containing 10 cells)}

\item{WithSpikes}{If true, the simulated dataset contains 20 spike-in genes}
}
\value{
An object of class \code{\link[BASiCS]{BASiCS_Data-class}}, simulated from the model implemented in BASiCS.
It contains 120 genes (100 biological and 20 spike-in) and 20 cells.
}
\description{
A simple \code{\link[BASiCS]{BASiCS_Data-class}} object is generated by simulating a dataset from the
model in BASiCS (Vallejos et al 2015). This is used for the examples in the package and the vignette.
}
\examples{
Data = makeExampleBASiCS_Data()
is(Data, "BASiCS_Data")

}
\references{
Vallejos, Marioni and Richardson (2015). Bayesian Analysis of Single-Cell Sequencing data.
}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}
}
