% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeExampleBASiCS_Data.R
\name{makeExampleBASiCS_Data}
\alias{makeExampleBASiCS_Data}
\title{Create a synthetic SingleCellExperiment example object with the 
format required for BASiCS}
\usage{
makeExampleBASiCS_Data(WithBatch = FALSE, WithSpikes = TRUE)
}
\arguments{
\item{WithBatch}{If \code{TRUE}, 2 batches are generated (each of them 
containing 10 cells). Default: \code{WithBatch = FALSE}.}

\item{WithSpikes}{If \code{TRUE}, the simulated dataset contains 20 spike-in 
genes. Default: \code{WithSpikes} \code{= TRUE}.}
}
\value{
An object of class \code{\linkS4class{SingleCellExperiment}}, with
synthetic data simulated from the model implemented in BASiCS.
If \code{WithSpikes = TRUE}, it contains 70 genes (50 biological and 
20 spike-in) and 20 cells. Alternatively, it contains 50 biological 
genes and 20 cells.
}
\description{
A synthetic \code{\linkS4class{SingleCellExperiment}} object is 
generated by simulating a dataset from the model underlying BASiCS. This is 
used to illustrate BASiCS in some of the package and vignette examples.
}
\examples{
Data <- makeExampleBASiCS_Data()
is(Data, 'SingleCellExperiment')

}
\references{
Vallejos, Marioni and Richardson (2015). PLoS Computational Biology. 

Vallejos, Marioni and Richardson (2016). Genome Biology.
}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl} and Nils Eling
}
