% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeExampleBASiCS_Data.R
\name{makeExampleBASiCS_Data}
\alias{makeExampleBASiCS_Data}
\title{Create a simple example of a SingleCellExperiment object with random data}
\usage{
makeExampleBASiCS_Data(WithBatch = FALSE, WithSpikes = TRUE, Example = 1)
}
\arguments{
\item{WithBatch}{If \code{TRUE}, 2 batches are generated (each of them containing 10 cells). Default value: \code{WithBatch = FALSE}.}

\item{WithSpikes}{If \code{TRUE}, the simulated dataset contains 20 spike-in genes. Default value: \code{WithSpikes = TRUE}.}

\item{Example}{Choice of example will be generated, defined by different parameter values (possible values = 1, 2). In both cases, expression counts associated to 50 biological genes are generated. 
This is used to generate two populations for differential testing. Default value \code{Example = 1}.}
}
\value{
An object of class \code{\link[SingleCellExperiment]{SingleCellExperiment}}, simulated from the model implemented in BASiCS.
If \code{WithSpikes = TRUE}, it contains 70 genes (50 biological and 20 spike-in) and 20 cells. Alternatively, it contains 50 biological genes and 20 cells.
}
\description{
A simple \code{\link[SingleCellExperiment]{SingleCellExperiment}} object is generated by simulating a dataset from the
model in BASiCS (Vallejos et al 2015). This is used to illustrate BASiCS in the package examples and in the vignette.
}
\examples{
Data = makeExampleBASiCS_Data()
is(Data, "SingleCellExperiment")

}
\references{
Vallejos, Marioni and Richardson (2015). Bayesian Analysis of Single-Cell Sequencing data. PLoS Computational Biology. 

Vallejos, Marioni and Richardson (2016). Beyond comparisons of means: understanding changes in gene expression at the single-cell level. Genome Biology.
}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl} and Nils Eling
}
