% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FunctionsPart1.R
\name{newBASiCS_Data}
\alias{newBASiCS_Data}
\title{Creates a BASiCS_Data object from a matrix of expression counts and experimental information about spike-in genes}
\usage{
newBASiCS_Data(Counts, Tech, SpikeInput, BatchInfo = NULL)
}
\arguments{
\item{Counts}{Matrix of dimensions \code{q} times \code{n} whose elements corresponds to the simulated expression counts.
First \code{q.bio} rows correspond to biological genes. Last \code{q-q.bio} rows correspond to technical spike-in genes.}

\item{Tech}{Logical vector of length \code{q}. If \code{Tech = F} the gene is biological; otherwise the gene is spike-in.}

\item{SpikeInput}{Vector of length \code{q-q.bio} whose elements indicate the simulated input concentrations for the spike-in genes.}

\item{BatchInfo}{Vector of length \code{n} whose elements indicate batch information.}
}
\value{
An object of class \code{\link[BASiCS]{BASiCS_Data-class}}.
}
\description{
\code{newBASiCS_Data} creates a \code{\link[BASiCS]{BASiCS_Data-class}} object from
a matrix of expression counts and experimental information about spike-in genes.
}
\examples{
set.seed(1)
Counts = matrix(rpois(10*5, 1), ncol = 5)
Tech = c(rep(FALSE,7),rep(TRUE,3))
set.seed(2)
SpikeInput = rgamma(3,1,1)
Data = newBASiCS_Data(Counts, Tech, SpikeInput)
}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}
\references{
Vallejos, Marioni and Richardson (2015). Bayesian Analysis of Single-Cell Sequencing data.
}
\seealso{
\code{\link[BASiCS]{BASiCS_Data-class}}
}

