% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsPart1.R
\name{newBASiCS_Data}
\alias{newBASiCS_Data}
\title{Creates a BASiCS_Data object from a matrix of expression counts and experimental information about spike-in genes}
\usage{
newBASiCS_Data(Counts, Tech, SpikeInfo, BatchInfo = NULL)
}
\arguments{
\item{Counts}{Matrix of dimensions \code{q} times \code{n} whose elements contain the expression counts to be analyses
(including biological and technical spike-in genes). Gene names must be stored as `rownames(Counts)`.}

\item{Tech}{Logical vector of length \code{q}. If \code{Tech = FALSE} the gene is biological; otherwise the gene is spike-in.}

\item{SpikeInfo}{\code{data.frame} whose first and second columns contain the gene names assigned to the spike-in genes
(they must match the ones in `rownames(Counts)`) and the associated input number of molecules, respectively.}

\item{BatchInfo}{Vector of length \code{n} whose elements indicate batch information.}
}
\value{
An object of class \code{\link[BASiCS]{BASiCS_Data-class}}.
}
\description{
\code{newBASiCS_Data} creates a \code{\link[BASiCS]{BASiCS_Data-class}} object from
a matrix of expression counts and experimental information about spike-in genes.
}
\examples{


# Expression counts
set.seed(1)
Counts = Counts = matrix(rpois(50*10, 2), ncol = 10)
rownames(Counts) <- c(paste0("Gene", 1:40), paste0("Spike", 1:10))

# Technical information
Tech = c(rep(FALSE,40),rep(TRUE,10))

# Spikes input number of molecules
set.seed(2)
SpikeInfo <- data.frame(gene=rownames(Counts)[Tech],amount=rgamma(10,1,1))

# Creating a BASiCS_Data object (no batch effect)
DataExample = newBASiCS_Data(Counts, Tech, SpikeInfo)

# Creating a BASiCS_Data object (with batch effect)
BatchInfo = c(rep(1, 5), rep(2, 5))
DataExample = newBASiCS_Data(Counts, Tech, SpikeInfo, BatchInfo)

# Thanks to Simon Andrews for reporting an issue in previous version of this documentation

}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
}
\references{
Vallejos, Marioni and Richardson (2015). Bayesian Analysis of Single-Cell Sequencing data.
}
\seealso{
\code{\link[BASiCS]{BASiCS_Data-class}}
}

