% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\docType{methods}
\name{plot-BASiCS_Chain-method}
\alias{plot-BASiCS_Chain-method}
\alias{plot}
\alias{plot,BASiCS_Chain-method}
\title{'plot' method for BASiCS_Chain objects}
\usage{
\S4method{plot}{BASiCS_Chain,ANY}(x, Param = "mu", Gene = NULL,
  Cell = NULL, Batch = 1, ylab = "", xlab = "", ...)
}
\arguments{
\item{x}{A \code{BASiCS_Chain} object.}

\item{Param}{Name of the slot to be used for the plot. Possible values: \code{mu, delta, phi, s, nu, theta}}

\item{Gene}{Specifies which gene is requested. Required only if \code{Param = "mu"} or \code{"delta"}}

\item{Cell}{Specifies which cell is requested. Required only if \code{Param = "phi", "s"} or \code{"nu"}}

\item{Batch}{Specifies which batch is requested. Required only if \code{Param = "theta"}}

\item{ylab}{As in \code{\link[graphics]{par}}.}

\item{xlab}{As in \code{\link[graphics]{par}}.}

\item{...}{Other graphical parameters (see \code{\link[graphics]{par}}).}
}
\description{
'plot' method for BASiCS_Chain objects
}
\examples{

# See
help(BASiCS_MCMC)

}
\references{
Vallejos, Marioni and Richardson (2015). Bayesian Analysis of Single-Cell Sequencing data.
}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}
}
