% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\docType{methods}
\name{plot-BASiCS_Chain-method}
\alias{plot-BASiCS_Chain-method}
\alias{plot,BASiCS_Chain-method}
\alias{plot,BASiCS_Chain,ANY-method}
\title{'plot' method for BASiCS_Chain objects}
\usage{
\S4method{plot}{BASiCS_Chain,ANY}(
  x,
  Param = "mu",
  Gene = NULL,
  Cell = NULL,
  Batch = 1,
  RegressionTerm = NULL,
  ylab = "",
  xlab = "",
  ...
)
}
\arguments{
\item{x}{A \code{\linkS4class{BASiCS_Chain}} object.}

\item{Param}{Name of the slot to be used for the plot.
Possible values: \code{'mu'}, \code{'delta'},
\code{'phi'}, \code{'s'}, \code{'nu'}, \code{'theta'},
\code{'beta'}, \code{'sigma2'} and \code{'epsilon'}.}

\item{Gene}{Specifies which gene is requested.
Required only if \code{Param = 'mu'} or \code{'delta'}}

\item{Cell}{Specifies which cell is requested.
Required only if \code{Param = 'phi', 's'} or \code{'nu'}}

\item{Batch}{Specifies which batch is requested.
Required only if \code{Param = 'theta'}}

\item{RegressionTerm}{Specifies which regression coefficient is requested.
Required only if \code{Param = 'beta'}}

\item{ylab}{As in \code{\link[graphics]{par}}.}

\item{xlab}{As in \code{\link[graphics]{par}}.}

\item{...}{Other graphical parameters (see \code{\link[graphics]{par}}).}
}
\value{
A plot object
}
\description{
'plot' method for \code{\linkS4class{BASiCS_Chain}} objects
}
\examples{

help(BASiCS_MCMC)

}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}

Nils Eling \email{eling@ebi.ac.uk}
}
