% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\docType{methods}
\name{plot-BASiCS_Summary-method}
\alias{plot-BASiCS_Summary-method}
\alias{plot,BASiCS_Summary-method,}
\alias{plot,BASiCS_Summary,ANY-method}
\title{'plot' method for BASiCS_Summary objects}
\usage{
\S4method{plot}{BASiCS_Summary,ANY}(x, Param = "mu", Param2 = NULL,
  Genes = NULL, Cells = NULL, Batches = NULL, xlab = "", ylab = "",
  xlim = "", ylim = "", pch = 16, col = "blue", bty = "n",
  SmoothPlot = TRUE, ...)
}
\arguments{
\item{x}{A \code{BASiCS_Summary} object.}

\item{Param}{Name of the slot to be used for the plot. Possible values: \code{mu, delta, phi, s, nu, theta}}

\item{Param2}{Name of the second slot to be used for the plot. Possible values: \code{mu, delta, phi, s, nu} (combinations between gene-specific and cell-specific parameters are not admitted)}

\item{Genes}{Specifies which genes are requested. Required only if \code{Param = "mu"} or \code{"delta"}}

\item{Cells}{Specifies which cells are requested. Required only if \code{Param = "phi", "s"} or \code{"nu"}}

\item{Batches}{Specifies which batches are requested. Required only if \code{Param = "theta"}}

\item{xlab}{As in \code{\link[graphics]{par}}.}

\item{ylab}{As in \code{\link[graphics]{par}}.}

\item{xlim}{As in \code{\link[graphics]{par}}.}

\item{ylim}{As in \code{\link[graphics]{par}}.}

\item{pch}{As in \code{\link[graphics]{par}}.}

\item{col}{As in \code{\link[graphics]{par}}.}

\item{bty}{As in \code{\link[graphics]{par}}.}

\item{SmoothPlot}{Logical if the \code{\link[graphics]{smoothScatter}} function should be used for plotting.}

\item{...}{Other graphical parameters (see \code{\link[graphics]{par}}).}
}
\description{
'plot' method for BASiCS_Summary objects
}
\examples{

# See
help(BASiCS_MCMC)

}
\references{
Vallejos, Marioni and Richardson (2015). Bayesian Analysis of Single-Cell Sequencing data.
}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}
}
