% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{rowData,BASiCS_ResultsDE-method}
\alias{rowData,BASiCS_ResultsDE-method}
\alias{rowData<-,BASiCS_ResultsDE-method}
\alias{rowData,BASiCS_ResultVG-method}
\alias{rowData<-,BASiCS_ResultVG-method}
\title{rowData getter and setter for \linkS4class{BASiCS_ResultsDE} 
and \linkS4class{BASiCS_ResultVG} objects.}
\usage{
\S4method{rowData}{BASiCS_ResultsDE}(x)

\S4method{rowData}{BASiCS_ResultsDE}(x) <- value

\S4method{rowData}{BASiCS_ResultVG}(x)

\S4method{rowData}{BASiCS_ResultVG}(x) <- value
}
\arguments{
\item{x}{\linkS4class{BASiCS_ResultVG} or \linkS4class{BASiCS_ResultsDE}
object.}

\item{value}{New \code{rowData} value for setter method.}
}
\value{
For the getter, a \linkS4class{DFrame}. For setter, the modified
\code{x}.
}
\description{
rowData getter and setter for \linkS4class{BASiCS_ResultsDE} 
and \linkS4class{BASiCS_ResultVG} objects.
}
