% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\docType{methods}
\name{Summary}
\alias{Summary}
\alias{Summary,BASiCS_Chain-method}
\title{'Summary' method for BASiCS_Chain objects}
\usage{
\S4method{Summary}{BASiCS_Chain}(x, prob = 0.95)
}
\arguments{
\item{x}{A \code{BASiCS_Chain} object.}

\item{prob}{\code{prob} argument for \code{\link[coda]{HPDinterval}} function.}
}
\value{
<<<<<<< HEAD
An object of class \code{\link[BASiCS]{BASiCS_Summary-class}}.
=======
An object of class \code{\link[BASiCS]{BASiCS_Summary}}.
>>>>>>> BioC
}
\description{
For each of the BASiCS parameters (see Vallejos et al 2015), 
\code{Summary} returns the corresponding postior medians and limits of the high posterior
density interval with probabilty equal to \code{prob}.
}
\examples{

<<<<<<< HEAD
Data = makeExampleBASiCS_Data()
MCMC_Output <- BASiCS_MCMC(Data, N = 50, Thin = 2, Burn = 2)
MCMC_Summary <- Summary(MCMC_Output)

# See documentation of function BASiCS_MCMC

}
\references{
Vallejos, Marioni and Richardson (2015). Bayesian Analysis of Single-Cell Sequencing data.
}
\author{
Catalina A. Vallejos \email{catalina.vallejos@mrc-bsu.cam.ac.uk}
=======
help(BASiCS_MCMC)

}
\references{
Vallejos, Marioni and Richardson (2015). Bayesian Analysis of Single-Cell Sequencing data. PLoS Computational Biology.
}
\author{
Catalina A. Vallejos \email{cnvallej@uc.cl}
>>>>>>> BioC
}
